refactor(dispatch): split API handlers by domain
