+++ src/app/dispatch.rs
@@ -5,6 +5,15 @@ use super::*;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::{mpsc::Sender, Arc};
 
+mod agents;
+mod content;
+mod core;
+mod extensions;
+mod orchestration;
+mod topology;
+
+type DispatchResult = Result<Value, (String, String)>;
+
 pub(crate) const MAX_AGENT_WAIT: Duration = Duration::from_secs(3600);
 pub(crate) const MAX_AGENT_WAITS_TOTAL: usize = 1024;
 pub(crate) const MAX_AGENT_WAITS_PER_PANE: usize = 64;
@@ -1152,3320 +1161,198 @@ impl App {
             ));
         }
         match method {
-            "ping" => Ok(json!({
-                "type":"pong",
-                "version": env!("CARGO_PKG_VERSION"),
-                "protocol":1,
-                "session": crate::session::display_name()
-            })),
-            "uhp.capabilities" => {
-                reject_api_fields(p, &[])?;
-                let mut capabilities =
-                    crate::api::capabilities(crate::ipc::api::current_sequence(&self.events));
-                if let Some(object) = capabilities.as_object_mut() {
-                    object.insert("session".into(), json!(crate::session::display_name()));
-                    object.insert(
-                        "server_generation".into(),
-                        json!(self.backend_server_generation),
-                    );
-                }
-                Ok(capabilities)
-            }
-            "config.get" => {
-                reject_api_fields(p, &[])?;
-                Ok(json!({"type":"config", "config":self.config}))
-            }
-            "config.patch" => {
-                reject_api_fields(p, &["patch"])?;
-                let patch = p.get("patch").ok_or_else(|| {
-                    (
-                        "invalid_request".to_string(),
-                        "config.patch needs a patch object".to_string(),
-                    )
-                })?;
-                let next = patched_config(&self.config, patch)?;
-                self.apply_socket_config(next, Some(patch))?;
-                Ok(json!({"type":"config", "config":self.config}))
-            }
-            "server.reload_config" => {
-                reject_api_fields(p, &[])?;
-                let next = crate::config::load();
-                self.apply_socket_config(next, None)?;
-                Ok(json!({"type":"config_reloaded", "config":self.config}))
-            }
-            "server.agent_manifests" => {
-                reject_api_fields(p, &[])?;
-                Ok(json!({
-                    "type":"agent_manifests",
-                    "rules":self.manifests.rule_count(),
-                    "agents":self.manifests.agent_names(),
-                }))
-            }
+            "ping" => self.api_ping(method, p),
+            "uhp.capabilities" => self.api_uhp_capabilities(method, p),
+            "config.get" => self.api_config_get(method, p),
+            "config.patch" => self.api_config_patch(method, p),
+            "server.reload_config" => self.api_server_reload_config(method, p),
+            "server.agent_manifests" => self.api_server_agent_manifests(method, p),
             "server.reload_agent_manifests" | "manifest.reload" => {
-                reject_api_fields(p, &[])?;
-                let manifests =
-                    crate::detect::Manifests::load(&crate::persist::ensure_manifests_dir());
-                self.apply_socket_manifests(manifests);
-                let rules = self.manifests.rule_count();
-                Ok(json!({"type":"agent_manifests_reloaded","rules":rules}))
+                self.api_server_reload_agent_manifests(method, p)
             }
-            "session.snapshot" => {
-                reject_api_fields(p, &[])?;
-                Ok(self.runtime_snapshot())
-            }
-            "search.capabilities" => Ok(json!({
-                "type": "search_capabilities",
-                "version": 1,
-                "methods": ["search.query", "search.activate"],
-                "scopes": ["all", "navigate", "files", "output"],
-                "max_results": crate::search::RESULT_CAP,
-                "max_response_bytes": crate::search::federation::MAX_SESSION_RESPONSE_BYTES,
-            })),
-            "theme.list" => Ok(self.theme_registry.list_json(&self.config.theme)),
-            "theme.path" => Ok(json!({
-                "type": "theme_path",
-                "path": crate::theme::themes_dir().display().to_string(),
-            })),
-            "theme.use" => {
-                let id = p.get("id").and_then(Value::as_str).ok_or_else(|| {
-                    (
-                        "invalid_request".to_string(),
-                        "theme.use needs an id".to_string(),
-                    )
-                })?;
-                if self.theme_registry.get(id).is_none() {
-                    return Err((
-                        "not_found".to_string(),
-                        format!("theme `{id}` is not installed"),
-                    ));
-                }
-                self.apply_theme(id);
-                Ok(json!({"type": "theme_selected", "id": self.config.theme}))
-            }
-            "server.stop" => {
-                self.should_quit = true;
-                Ok(json!({"type":"ok"}))
-            }
-            "pane.get" => {
-                reject_api_fields(p, &["pane"])?;
-                let pane = self.resolve_pane(p)?.ok_or_else(not_found)?;
-                self.socket_pane(pane)
-            }
-            "pane.current" => {
-                reject_api_fields(p, &[])?;
-                self.socket_pane(self.layout().focus)
-            }
-            "pane.layout" => {
-                reject_api_fields(p, &["pane"])?;
-                let pane = self.resolve_pane_or_focus(p)?;
-                self.socket_pane_layout(pane)
-            }
-            "pane.neighbor" => {
-                reject_api_fields(p, &["pane", "direction"])?;
-                let pane = self.resolve_pane_or_focus(p)?;
-                let direction = crate::api::topology::direction(p)?;
-                let (workspace, tab) = self.pane_location(pane).ok_or_else(not_found)?;
-                let layout = &self.workspaces[workspace].tabs[tab].layout;
-                let neighbor =
-                    layout.neighbor(crate::api::topology::logical_area(), pane, direction);
-                Ok(json!({"type":"pane_neighbor","pane":pane.0.to_string(),
-                    "neighbor":neighbor.map(|id| id.0.to_string())}))
-            }
-            "pane.edges" => {
-                reject_api_fields(p, &["pane"])?;
-                let pane = self.resolve_pane_or_focus(p)?;
-                let (workspace, tab) = self.pane_location(pane).ok_or_else(not_found)?;
-                let area = crate::api::topology::logical_area();
-                let rect = self.workspaces[workspace].tabs[tab]
-                    .layout
-                    .pane_rect(area, pane)
-                    .ok_or_else(not_found)?;
-                Ok(
-                    json!({"type":"pane_edges","pane":pane.0.to_string(),"edges":{
-                        "left":rect.x == area.x, "right":rect.right() == area.right(),
-                        "top":rect.y == area.y, "bottom":rect.bottom() == area.bottom(),
-                    }}),
-                )
-            }
-            "pane.list" => {
-                let focus = self.layout().focus;
-                let panes: Vec<Value> = self
-                    .layout()
-                    .leaves()
-                    .iter()
-                    .map(|id| {
-                        let (agent, status) = self
-                            .status
-                            .get(id)
-                            .map(|s| (s.agent.clone(), state_str(s.state).to_string()))
-                            .unwrap_or_else(|| (String::new(), "unknown".to_string()));
-                        let cwd = self
-                            .panes
-                            .get(id)
-                            .map(|p| p.cwd.display().to_string())
-                            .unwrap_or_default();
-                        let history = self.panes.get(id).map(|p| p.history_metrics());
-                        let module = self
-                            .module_panes
-                            .get(id)
-                            .map(|r| json!({"id": r.module_id, "entrypoint": r.entrypoint}));
-                        json!({
-                            "pane": id.0.to_string(), "agent": agent, "status": status,
-                            "focused": *id == focus, "cwd": cwd, "module": module,
-                            "scroll_offset": history.map(|m| m.offset).unwrap_or(0),
-                            "history_rows": history.map(|m| m.retained_rows).unwrap_or(0),
-                            "history_budget_bytes": history.map(|m| m.budget_bytes).unwrap_or(0),
-                            "history_bytes": history.map(|m| m.retained_bytes).unwrap_or(0),
-                            "history_estimated_grid_bytes": history.map(|m| m.estimated_grid_bytes).unwrap_or(0),
-                            "history_cache_bytes": history.and_then(|m| m.cache_bytes),
-                            "history_compacted_rows": history.and_then(|m| m.compacted_rows),
-                            "history_allocated_cells": history.and_then(|m| m.allocated_cells),
-                            "history_packed_blocks": history.and_then(|m| m.packed_blocks),
-                            "history_packed_bytes": history.and_then(|m| m.packed_bytes),
-                            "history_packed_rows": history.and_then(|m| m.packed_rows),
-                            "history_dense_row_bytes": history.and_then(|m| m.dense_row_bytes),
-                            "history_row_descriptor_bytes": history.and_then(|m| m.row_descriptor_bytes),
-                            "history_allocation_count": history.and_then(|m| m.allocation_count),
-                            "history_exact": history.map(|m| m.exact_bytes).unwrap_or(false),
-                            "history_bytes_kind": if history.is_some_and(|m| m.exact_bytes) { "exact" } else { "estimated" },
-                        })
-                    })
-                    .collect();
-                Ok(json!({
-                    "type":"pane_list",
-                    "panes":panes,
-                    "detection_extractions": self.detection_extractions,
-                    "detection_skips": self.detection_skips,
-                    "detection_performance": {
-                        "panes_considered": self.detection_panes_considered,
-                        "panes_extracted": self.detection_extractions,
-                        "panes_generation_skipped": self.detection_skips,
-                        "full_fleet_audits": self.detection_full_fleet_audits,
-                        "audit_recoveries": self.detection_audit_recoveries,
-                    },
-                    "render_performance": crate::ipc::server::performance_snapshot(),
-                }))
-            }
-            "pane.split" => {
-                if self.workspaces.is_empty() && !self.ensure_workspace_for_terminal() {
-                    return Err((
-                        "spawn_failed".to_string(),
-                        "could not create a terminal in the home directory".to_string(),
-                    ));
-                }
-                let base = match p.get("pane") {
-                    None | Some(Value::Null) => self.layout().focus,
-                    Some(_) => self.resolve_pane(p)?.ok_or_else(not_found)?,
-                };
-                let dir = p
-                    .get("direction")
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("right");
-                let axis = if dir == "down" || dir == "stack" {
-                    Axis::Row
-                } else {
-                    Axis::Col
-                };
-                let focus = p.get("focus").and_then(|v| v.as_bool()) != Some(false);
-                let new = self.split_pane(base, axis, focus).ok_or_else(not_found)?;
-                let (workspace, tab) = self.pane_location(new).ok_or_else(not_found)?;
-                Ok(json!({
-                    "type":"pane",
-                    "pane": new.0.to_string(),
-                    "workspace": workspace.to_string(),
-                    "tab": (tab + 1).to_string(),
-                }))
-            }
-            "pane.move" => {
-                let id = self.resolve_pane(p)?.ok_or_else(not_found)?;
-                let new_tab = match p.get("new_tab") {
-                    None => false,
-                    Some(Value::Bool(v)) => *v,
-                    Some(_) => {
-                        return Err((
-                            "invalid_request".to_string(),
-                            "new_tab must be a boolean".to_string(),
-                        ))
-                    }
-                };
-                let tab = param_usize(p, "tab");
-                if new_tab == tab.is_some() {
-                    return Err((
-                        "invalid_request".to_string(),
-                        "pass exactly one destination: tab (1-based) or new_tab=true".to_string(),
-                    ));
-                }
-                let target = if new_tab {
-                    MoveTarget::NewTab
-                } else {
-                    MoveTarget::Tab(required_one_based_param(p, "tab")?)
-                };
-                let moved = self.move_pane_to_tab(id, target).map_err(pane_move_error)?;
-                Ok(json!({
-                    "type": "pane_move",
-                    "pane": id.0.to_string(),
-                    "workspace": moved.workspace.to_string(),
-                    "tab": (moved.tab + 1).to_string(),
-                }))
-            }
-            "pane.run" => {
-                let id = self.resolve_pane(p)?.ok_or_else(not_found)?;
-                let cmd = p.get("command").and_then(|v| v.as_str()).unwrap_or("");
-                let pane = self.panes.get(&id).ok_or_else(not_found)?;
-                let mut bytes = Vec::with_capacity(cmd.len() + 1);
-                bytes.extend_from_slice(cmd.as_bytes());
-                bytes.push(b'\r');
-                pane.try_send(&bytes)
-                    .map_err(|message| ("send_failed".to_string(), message))?;
-                Ok(json!({"type":"ok"}))
-            }
-            "pane.send_input" => {
-                let id = self.resolve_pane(p)?.ok_or_else(not_found)?;
-                let text = p.get("text").and_then(|v| v.as_str()).unwrap_or("");
-                let paste = match p.get("paste") {
-                    None => false,
-                    Some(Value::Bool(paste)) => *paste,
-                    Some(_) => {
-                        return Err((
-                            "invalid_request".to_string(),
-                            "paste must be a boolean".to_string(),
-                        ))
-                    }
-                };
-                let pane = self.panes.get(&id).ok_or_else(not_found)?;
-                let result = if paste {
-                    pane.try_send_paste(text)
-                } else {
-                    pane.try_send(text.as_bytes())
-                };
-                result.map_err(|message| ("send_failed".to_string(), message))?;
-                Ok(json!({"type":"ok"}))
-            }
-            "pane.read" => {
-                let id = self.resolve_pane(p)?.ok_or_else(not_found)?;
-                let lines = p.get("lines").and_then(|v| v.as_u64()).unwrap_or(200) as u16;
-                let text = self
-                    .panes
-                    .get(&id)
-                    .and_then(|pane| pane.engine.lock().ok().map(|e| e.detection_text(lines)))
-                    .unwrap_or_default();
-                Ok(json!({"type":"pane_read","text":text}))
-            }
+            "session.snapshot" => self.api_session_snapshot(method, p),
+            "search.capabilities" => self.api_search_capabilities(method, p),
+            "theme.list" => self.api_theme_list(method, p),
+            "theme.path" => self.api_theme_path(method, p),
+            "theme.use" => self.api_theme_use(method, p),
+            "server.stop" => self.api_server_stop(method, p),
+            "pane.get" => self.api_pane_get(method, p),
+            "pane.current" => self.api_pane_current(method, p),
+            "pane.layout" => self.api_pane_layout(method, p),
+            "pane.neighbor" => self.api_pane_neighbor(method, p),
+            "pane.edges" => self.api_pane_edges(method, p),
+            "pane.list" => self.api_pane_list(method, p),
+            "pane.split" => self.api_pane_split(method, p),
+            "pane.move" => self.api_pane_move(method, p),
+            "pane.run" => self.api_pane_run(method, p),
+            "pane.send_input" => self.api_pane_send_input(method, p),
+            "pane.read" => self.api_pane_read(method, p),
             // Global scrollback search (docs/63): scan every pane's retained
             // output. Returns matches with the scroll offset that lands on each,
             // plus the total found (which may exceed the returned, capped, list).
-            "search" => {
-                let query = p.get("query").and_then(|v| v.as_str()).unwrap_or("").trim();
-                let case_sensitive = p
-                    .get("case_sensitive")
-                    .and_then(|v| v.as_bool())
-                    .unwrap_or(false);
-                let (hits, total) = self.search_all(query, case_sensitive);
-                let matches: Vec<Value> = hits
-                    .iter()
-                    .map(|h| {
-                        json!({
-                            "pane": h.pane.0.to_string(),
-                            "workspace": h.ws,
-                            "workspace_name": h.ws_name,
-                            "line_offset": h.offset,
-                            "text": h.line,
-                            "col": h.col,
-                        })
-                    })
-                    .collect();
-                Ok(json!({
-                    "type": "search",
-                    "query": query,
-                    "total": total,
-                    "shown": matches.len(),
-                    "matches": matches,
-                }))
-            }
-            "pane.close" => {
-                let id = self.resolve_pane(p)?.ok_or_else(not_found)?;
-                self.close_pane(id);
-                Ok(json!({"type":"ok"}))
-            }
+            "search" => self.api_search(method, p),
+            "pane.close" => self.api_pane_close(method, p),
             // A **global** single-pane status lookup (any workspace) — `pane.list` is
             // scoped to the active workspace, so `luvus wait agent-status` polls this.
-            "pane.status" => {
-                let id = self.resolve_pane(p)?.ok_or_else(not_found)?;
-                let (agent, status, authority, state_source) = self
-                    .status
-                    .get(&id)
-                    .map(|s| {
-                        (
-                            s.agent.clone(),
-                            state_str(s.state).to_string(),
-                            s.identity_source,
-                            s.state_source,
-                        )
-                    })
-                    .unwrap_or_else(|| (String::new(), "unknown".to_string(), "none", "none"));
-                let history = self.panes.get(&id).map(|p| p.history_metrics());
-                Ok(json!({
-                    "type":"pane_status","pane": id.0.to_string(), "agent": agent, "status": status,
-                    "authority":authority, "state_source":state_source,
-                    "scroll_offset": history.map(|m| m.offset).unwrap_or(0),
-                    "history_rows": history.map(|m| m.retained_rows).unwrap_or(0),
-                    "history_budget_bytes": history.map(|m| m.budget_bytes).unwrap_or(0),
-                    "history_bytes": history.map(|m| m.retained_bytes).unwrap_or(0),
-                    "history_estimated_grid_bytes": history.map(|m| m.estimated_grid_bytes).unwrap_or(0),
-                    "history_cache_bytes": history.and_then(|m| m.cache_bytes),
-                    "history_compacted_rows": history.and_then(|m| m.compacted_rows),
-                    "history_allocated_cells": history.and_then(|m| m.allocated_cells),
-                    "history_packed_blocks": history.and_then(|m| m.packed_blocks),
-                    "history_packed_bytes": history.and_then(|m| m.packed_bytes),
-                    "history_packed_rows": history.and_then(|m| m.packed_rows),
-                    "history_dense_row_bytes": history.and_then(|m| m.dense_row_bytes),
-                    "history_row_descriptor_bytes": history.and_then(|m| m.row_descriptor_bytes),
-                    "history_allocation_count": history.and_then(|m| m.allocation_count),
-                    "history_exact": history.map(|m| m.exact_bytes).unwrap_or(false),
-                    "history_bytes_kind": if history.is_some_and(|m| m.exact_bytes) { "exact" } else { "estimated" },
-                }))
-            }
-            "pane.processes" => {
-                reject_api_fields(p, &["pane"])?;
-                let id = self.resolve_pane(p)?.ok_or_else(not_found)?;
-                self.request_proc_scan_if_stale(id);
-                Ok(self.pane_processes(id))
-            }
-            "pane.report_session" => {
-                reject_api_fields(p, &["pane", "agent", "session_id", "usage"])?;
-                let id = self.resolve_pane(p)?.ok_or_else(not_found)?;
-                let raw_agent = required_bounded_string(p, "agent", 64)?;
-                let agent = crate::agent::canonical_builtin(&raw_agent).ok_or_else(|| {
-                    (
-                        "invalid_request".to_string(),
-                        "agent must name a built-in Luvus adapter".to_string(),
-                    )
-                })?;
-                let session_id = required_bounded_string(p, "session_id", 256)?;
-                if !crate::agent::safe_session_id(&session_id) {
-                    return Err((
-                        "invalid_request".to_string(),
-                        "session_id must contain only safe identifier characters".to_string(),
-                    ));
-                }
-
-                let key = crate::mission::UsageKey::new(agent, &session_id);
-                let usage = p.get("usage").map(parse_reported_usage).transpose()?;
-                self.prune_reported_usage();
-
-                if self.status.iter().any(|(pane, status)| {
-                    *pane != id
-                        && status.agent_session.as_ref().is_some_and(|session| {
-                            session.agent == key.agent && session.session_id == key.session_id
-                        })
-                }) {
-                    return Err((
-                        "conflict".to_string(),
-                        "session is already owned by another pane".to_string(),
-                    ));
-                }
-                if let Some(report) = &usage {
-                    if let Some(previous) = self.reported_usage.get(&key) {
-                        if previous.pane != id {
-                            return Err((
-                                "conflict".to_string(),
-                                "session usage is already owned by another pane".to_string(),
-                            ));
-                        }
-                        if report.updated_at < previous.updated_at {
-                            return Err((
-                                "stale_report".to_string(),
-                                "usage report is older than the current value".to_string(),
-                            ));
-                        }
-                    }
-                }
-
-                let replaced = self
-                    .reported_usage
-                    .iter()
-                    .filter_map(|(existing, owner)| {
-                        (owner.pane == id && existing != &key).then_some(existing.clone())
-                    })
-                    .collect::<Vec<_>>();
-                if usage.is_some()
-                    && !reported_usage_has_capacity(
-                        self.reported_usage.len(),
-                        self.reported_usage.contains_key(&key),
-                        !replaced.is_empty(),
-                    )
-                {
-                    return Err((
-                        "resource_exhausted".to_string(),
-                        "reported usage cache is full".to_string(),
-                    ));
-                }
-                for existing in replaced {
-                    self.reported_usage.remove(&existing);
-                    self.agent_usage.remove(&existing);
-                    self.usage_mtimes.remove(&existing);
-                }
-
-                let status = self.status.get_mut(&id).ok_or_else(not_found)?;
-                status.agent = agent.to_string();
-                status.agent_session = Some(AgentSession {
-                    agent: agent.to_string(),
-                    session_id,
-                });
-                status.force_detect = true;
-
-                if let Some(mut report) = usage {
-                    if !self.config.mission_pricing.is_empty() {
-                        report.value.cost = crate::mission::estimate_cost_with(
-                            &report.value.model,
-                            report.value.tokens_in,
-                            report.value.tokens_out,
-                            report.value.cache,
-                            &self.config.mission_pricing,
-                        );
-                    }
-                    self.agent_usage.insert(key.clone(), report.value);
-                    self.reported_usage.insert(
-                        key.clone(),
-                        crate::mission::ReportedUsage {
-                            pane: id,
-                            updated_at: report.updated_at,
-                        },
-                    );
-                    self.usage_mtimes.remove(&key);
-                }
-                self.session_dirty = true;
-                self.confirm_durable_active_target(id);
-                Ok(json!({"type":"ok"}))
-            }
+            "pane.status" => self.api_pane_status(method, p),
+            "pane.processes" => self.api_pane_processes(method, p),
+            "pane.report_session" => self.api_pane_report_session(method, p),
             // A precise agent lifecycle event from an integration hook:
             // permission prompt, question, turn end. Forwarded verbatim onto the
             // event bus as `agent.hook` for modules and API clients.
-            "pane.report_event" => {
-                let id = self.resolve_pane(p)?.ok_or_else(not_found)?;
-                let agent = p.get("agent").and_then(|v| v.as_str()).unwrap_or("");
-                let kind = p.get("kind").and_then(|v| v.as_str()).unwrap_or("");
-                let message = p.get("message").and_then(|v| v.as_str()).unwrap_or("");
-                let tool = p.get("tool").and_then(|v| v.as_str()).unwrap_or("");
-                self.emit_event(
-                    "agent.hook",
-                    json!({ "pane": id.0.to_string(), "agent": agent, "kind": kind, "message": message, "tool": tool }),
-                );
-                Ok(json!({"type":"ok"}))
-            }
+            "pane.report_event" => self.api_pane_report_event(method, p),
             // ── workspaces ── (`node.*` kept as a back-compat alias)
-            "workspace.list" | "node.list" => {
-                let active = self.active_ws;
-                let mut display_positions = vec![0usize; self.workspaces.len()];
-                for (position, (workspace, _)) in
-                    self.workspace_display_order().into_iter().enumerate()
-                {
-                    display_positions[workspace] = position;
-                }
-                let arr: Vec<Value> = self
-                    .workspaces
-                    .iter()
-                    .enumerate()
-                    .map(|(i, w)| {
-                        let terminal_cwd = self
-                            .workspace_terminal_cwd(i)
-                            .unwrap_or(&w.cwd)
-                            .display()
-                            .to_string();
-                        json!({
-                            "workspace": i.to_string(),
-                            "workspace_id": w.id,
-                            "name": w.name,
-                            "cwd": w.cwd.display().to_string(),
-                            "terminal_cwd": terminal_cwd,
-                            "pinned": w.pinned,
-                            "display_position": display_positions[i].to_string(),
-                            "active": i == active,
-                            "tabs": w.tabs.len(),
-                        })
-                    })
-                    .collect();
-                Ok(json!({"type":"workspace_list","workspaces":arr}))
-            }
-            "workspace.get" => {
-                reject_api_fields(p, &["workspace", "workspace_id"])?;
-                let index = self.optional_socket_workspace(p)?.unwrap_or(self.active_ws);
-                self.socket_workspace(index)
-            }
-            "workspace.move" => {
-                reject_api_fields(p, &["workspace", "workspace_id", "to"])?;
-                let workspace = self.required_socket_workspace(p)?;
-                let to = required_index_param(p, "to")?;
-                let positions = self.reorder_workspace_block(&[workspace], to)?;
-                self.emit_event(
-                    "workspace.moved",
-                    json!({"workspace":workspace.to_string(),"to":positions[0].to_string()}),
-                );
-                Ok(json!({
-                    "type":"workspace_move",
-                    "workspace":workspace.to_string(),
-                    "to":positions[0].to_string()
-                }))
-            }
-            "workspace.move_block" => {
-                reject_api_fields(p, &["workspaces", "workspace_ids", "to"])?;
-                if p.get("workspaces").is_some() && p.get("workspace_ids").is_some() {
-                    return Err((
-                        "invalid_request".to_string(),
-                        "workspaces and workspace_ids cannot be used together".to_string(),
-                    ));
-                }
-                let values = p
-                    .get("workspaces")
-                    .or_else(|| p.get("workspace_ids"))
-                    .and_then(Value::as_array)
-                    .ok_or_else(|| {
-                        (
-                            "invalid_request".to_string(),
-                            "workspaces or workspace_ids must be an array".to_string(),
-                        )
-                    })?;
-                if values.is_empty()
-                    || values.len() > crate::api::topology::MAX_WORKSPACE_MOVE_BLOCK
-                {
-                    return Err((
-                        "limit_exceeded".to_string(),
-                        "workspace block size is invalid".to_string(),
-                    ));
-                }
-                let workspaces: Vec<usize> = if p.get("workspace_ids").is_some() {
-                    values
-                        .iter()
-                        .map(|value| {
-                            let id = value.as_str().ok_or_else(|| {
-                                (
-                                    "invalid_request".to_string(),
-                                    "workspace_ids must contain strings".to_string(),
-                                )
-                            })?;
-                            self.workspaces
-                                .iter()
-                                .position(|workspace| workspace.id == id)
-                                .ok_or_else(|| {
-                                    (
-                                        "not_found".to_string(),
-                                        format!("workspace id {id} not found"),
-                                    )
-                                })
-                        })
-                        .collect::<Result<_, _>>()?
-                } else {
-                    values
-                        .iter()
-                        .map(parse_index_value)
-                        .collect::<Result<_, _>>()?
-                };
-                let to = required_index_param(p, "to")?;
-                let positions = self.reorder_workspace_block(&workspaces, to)?;
-                self.emit_event(
-                    "workspace.block_moved",
-                    json!({"workspaces":workspaces,"positions":positions}),
-                );
-                Ok(json!({
-                    "type":"workspace_move_block",
-                    "workspaces":workspaces,
-                    "positions":positions
-                }))
-            }
-            "workspace.report_metadata" => {
-                reject_api_fields(
-                    p,
-                    &["workspace", "workspace_id", "branch", "ahead", "behind"],
-                )?;
-                let index = self.required_socket_workspace(p)?;
-                let branch = optional_nullable_string(p, "branch", 512)?;
-                let ahead = optional_u32(p, "ahead")?;
-                let behind = optional_u32(p, "behind")?;
-                let workspace = self
-                    .workspaces
-                    .get_mut(index)
-                    .ok_or_else(|| workspace_update_error(index, WorkspaceUpdateError::NotFound))?;
-                if p.get("branch").is_some() {
-                    workspace.branch = branch;
-                }
-                if ahead.is_some() || behind.is_some() {
-                    let current = workspace.git_ahead_behind.unwrap_or_default();
-                    workspace.git_ahead_behind =
-                        Some((ahead.unwrap_or(current.0), behind.unwrap_or(current.1)));
-                }
-                self.emit_event(
-                    "workspace.metadata_reported",
-                    json!({"workspace":index.to_string()}),
-                );
-                self.socket_workspace(index)
-            }
-            "workspace.new" | "node.new" => {
-                self.new_workspace();
-                Ok(json!({
-                    "type":"workspace",
-                    "workspace": self.active_ws.to_string()
-                }))
-            }
-            "workspace.open" | "node.open" => {
-                // Open `path` as a workspace, or focus it if it's already one. Used
-                // when `luvus` attaches to a running server from a new folder, so the
-                // launch directory shows up as a workspace.
-                //
-                // `focus` (default true) governs the *already-open* case. The
-                // automatic attach-open (`open_cwd_workspace`) passes `false`: it
-                // ensures the launch folder is a workspace but must NOT steal focus
-                // from the workspace a restored session left you on — otherwise
-                // reopening `luvus` always snaps back to the launch folder (usually
-                // the first workspace), never the one you were last using. An
-                // explicit `luvus workspace open <path>` omits it and still focuses.
-                let path = PathBuf::from(req_str(p, "path")?);
-                let focus = p.get("focus").and_then(|v| v.as_bool()).unwrap_or(true);
-                match self
-                    .workspaces
-                    .iter()
-                    .position(|w| crate::platform::same_path(&w.cwd, &path))
-                {
-                    Some(i) => {
-                        self.forget_closed_workspace_path(&path);
-                        if focus {
-                            self.active_ws = i;
-                        }
-                    }
-                    None if !focus && self.automatic_workspace_open_is_suppressed(&path) => {}
-                    // Report a failed open instead of answering with the
-                    // *previously* active node, which read as success and left
-                    // the caller (and the user) looking at the wrong folder.
-                    None if !self.create_workspace_at(path.clone()) => {
-                        return Err((
-                            "spawn_failed".to_string(),
-                            format!(
-                                "couldn't open {} — the shell failed to start there",
-                                path.display()
-                            ),
-                        ));
-                    }
-                    None => {}
-                }
-                Ok(json!({
-                    "type":"workspace",
-                    "workspace": self.active_ws.to_string()
-                }))
-            }
-            "workspace.focus" | "node.focus" => {
-                if let Some(i) = self.optional_socket_workspace(p)? {
-                    if i < self.workspaces.len() {
-                        self.active_ws = i;
-                    }
-                }
-                Ok(json!({"type":"ok"}))
-            }
-            "workspace.rename" | "node.rename" => {
-                let i = self.required_socket_workspace(p)?;
-                let name = p.get("name").and_then(|v| v.as_str()).ok_or_else(|| {
-                    (
-                        "invalid_request".to_string(),
-                        "name must be a non-empty string".to_string(),
-                    )
-                })?;
-                self.rename_workspace(i, name)
-                    .map_err(|err| workspace_update_error(i, err))?;
-                let workspace = &self.workspaces[i];
-                Ok(json!({
-                    "type": "workspace_rename",
-                    "workspace": i.to_string(),
-                    "name": workspace.name,
-                    "cwd": workspace.cwd.display().to_string(),
-                    "pinned": workspace.pinned,
-                    "display_position": self.workspace_display_position(i).unwrap_or(i).to_string(),
-                }))
-            }
-            "workspace.pin" | "node.pin" => {
-                let i = self.required_socket_workspace(p)?;
-                let pinned = p.get("pinned").and_then(|v| v.as_bool()).ok_or_else(|| {
-                    (
-                        "invalid_request".to_string(),
-                        "pinned must be a boolean".to_string(),
-                    )
-                })?;
-                self.set_workspace_pinned(i, pinned)
-                    .map_err(|err| workspace_update_error(i, err))?;
-                let workspace = &self.workspaces[i];
-                Ok(json!({
-                    "type": "workspace_pin",
-                    "workspace": i.to_string(),
-                    "name": workspace.name,
-                    "cwd": workspace.cwd.display().to_string(),
-                    "pinned": workspace.pinned,
-                    "display_position": self.workspace_display_position(i).unwrap_or(i).to_string(),
-                }))
-            }
-            "workspace.close" | "node.close" => {
-                let i = self.optional_socket_workspace(p)?.unwrap_or(self.active_ws);
-                self.close_workspace(i);
-                Ok(json!({"type":"ok"}))
-            }
+            "workspace.list" | "node.list" => self.api_workspace_list(method, p),
+            "workspace.get" => self.api_workspace_get(method, p),
+            "workspace.move" => self.api_workspace_move(method, p),
+            "workspace.move_block" => self.api_workspace_move_block(method, p),
+            "workspace.report_metadata" => self.api_workspace_report_metadata(method, p),
+            "workspace.new" | "node.new" => self.api_workspace_new(method, p),
+            "workspace.open" | "node.open" => self.api_workspace_open(method, p),
+            "workspace.focus" | "node.focus" => self.api_workspace_focus(method, p),
+            "workspace.rename" | "node.rename" => self.api_workspace_rename(method, p),
+            "workspace.pin" | "node.pin" => self.api_workspace_pin(method, p),
+            "workspace.close" | "node.close" => self.api_workspace_close(method, p),
             // ── tabs ──
-            "tab.list" => {
-                let ws = self.ws();
-                let arr: Vec<Value> = ws
-                    .tabs
-                    .iter()
-                    .enumerate()
-                    .map(|(i, t)| {
-                        // `name` is what `tab.rename` writes; `kind` distinguishes
-                        // the dashboard tabs, which have no panes and can't be named.
-                        let kind = if t.git.is_some() {
-                            "git"
-                        } else if t.orch {
-                            "orch"
-                        } else {
-                            "panes"
-                        };
-                        json!({
-                            "tab": (i + 1).to_string(),
-                            "tab_id": t.id,
-                            "active": i == ws.active_tab,
-                            "name": t.name.clone(),
-                            "kind": kind,
-                        })
-                    })
-                    .collect();
-                Ok(json!({"type":"tab_list","tabs":arr}))
-            }
-            "tab.get" => {
-                reject_api_fields(p, &["workspace", "workspace_id", "tab", "tab_id"])?;
-                let workspace = self.optional_socket_workspace(p)?.unwrap_or(self.active_ws);
-                let tab = self
-                    .optional_socket_tab(workspace, p, "tab", "tab_id")?
-                    .unwrap_or_else(|| {
-                        self.workspaces
-                            .get(workspace)
-                            .map(|ws| ws.active_tab)
-                            .unwrap_or(usize::MAX)
-                    });
-                self.socket_tab(workspace, tab)
-            }
-            "tab.new" => {
-                if self.workspaces.is_empty() {
-                    if !self.ensure_workspace_for_terminal() {
-                        return Err((
-                            "spawn_failed".to_string(),
-                            "could not create a terminal in the home directory".to_string(),
-                        ));
-                    }
-                } else {
-                    self.new_tab();
-                }
-                Ok(json!({
-                    "type":"tab",
-                    "tab": (self.ws().active_tab + 1).to_string()
-                }))
-            }
-            "tab.focus" => {
-                let index = self.required_socket_tab(self.active_ws, p, "tab", "tab_id")?;
-                self.focus_tab(index).map_err(tab_focus_error)?;
-                Ok(json!({"type":"ok"}))
-            }
-            "tab.move" => {
-                let (from, to, active) = if let Some(raw_direction) =
-                    p.get("direction").filter(|direction| !direction.is_null())
-                {
-                    if p.get("to").is_some() {
-                        return Err((
-                            "invalid_request".to_string(),
-                            "direction and to cannot be used together".to_string(),
-                        ));
-                    }
-                    let direction = match raw_direction.as_str() {
-                        Some("left") => TabMoveDirection::Left,
-                        Some("right") => TabMoveDirection::Right,
-                        _ => {
-                            return Err((
-                                "invalid_request".to_string(),
-                                "direction must be left or right".to_string(),
-                            ))
-                        }
-                    };
-                    let from = self.optional_socket_tab(self.active_ws, p, "tab", "tab_id")?;
-                    self.move_tab_direction(from, direction)
-                        .map_err(tab_move_error)?
-                } else {
-                    let from = self.required_socket_tab(self.active_ws, p, "tab", "tab_id")?;
-                    let to = required_one_based_param(p, "to")?;
-                    let active = self.move_tab(from, to).map_err(tab_move_error)?;
-                    (from, to, active)
-                };
-                Ok(json!({
-                    "type": "tab_move",
-                    "from": (from + 1).to_string(),
-                    "to": (to + 1).to_string(),
-                    "active": (active + 1).to_string(),
-                }))
-            }
-            "tab.swap" => {
-                let first = self.required_socket_tab(self.active_ws, p, "tab", "tab_id")?;
-                let second = self.required_socket_tab(self.active_ws, p, "with", "with_id")?;
-                let active = self.swap_tabs(first, second).map_err(tab_move_error)?;
-                Ok(json!({
-                    "type": "tab_swap",
-                    "tab": (first + 1).to_string(),
-                    "with": (second + 1).to_string(),
-                    "active": (active + 1).to_string(),
-                }))
-            }
+            "tab.list" => self.api_tab_list(method, p),
+            "tab.get" => self.api_tab_get(method, p),
+            "tab.new" => self.api_tab_new(method, p),
+            "tab.focus" => self.api_tab_focus(method, p),
+            "tab.move" => self.api_tab_move(method, p),
+            "tab.swap" => self.api_tab_swap(method, p),
             // Name a tab from a module (docs/13 §3.9) — the same label the
             // tab-rename modal writes. An empty name clears it back to a number.
-            "tab.rename" => {
-                let index = self
-                    .optional_socket_tab(self.active_ws, p, "tab", "tab_id")?
-                    .unwrap_or(self.ws().active_tab);
-                let name = p.get("name").and_then(Value::as_str).ok_or_else(|| {
-                    (
-                        "invalid_request".to_string(),
-                        "name must be a string (empty clears the tab name)".to_string(),
-                    )
-                })?;
-                self.rename_tab(index, name).map_err(tab_rename_error)?;
-                Ok(json!({"type":"ok"}))
-            }
-            "tab.close" => {
-                let i = self
-                    .optional_socket_tab(self.active_ws, p, "tab", "tab_id")?
-                    .unwrap_or(self.ws().active_tab);
-                self.close_tab(i);
-                Ok(json!({"type":"ok"}))
-            }
-            "layout.export" => {
-                reject_api_fields(p, &["workspace", "workspace_id", "tab", "tab_id"])?;
-                let workspace = self.optional_socket_workspace(p)?.unwrap_or(self.active_ws);
-                let tab = self
-                    .optional_socket_tab(workspace, p, "tab", "tab_id")?
-                    .unwrap_or_else(|| {
-                        self.workspaces
-                            .get(workspace)
-                            .map(|ws| ws.active_tab)
-                            .unwrap_or(usize::MAX)
-                    });
-                let tab_ref = self
-                    .workspaces
-                    .get(workspace)
-                    .and_then(|ws| ws.tabs.get(tab))
-                    .ok_or_else(not_found)?;
-                if !tab_ref.is_renameable() {
-                    return Err((
-                        "invalid_request".to_string(),
-                        "dashboard tabs do not have a mutable pane layout".to_string(),
-                    ));
-                }
-                Ok(
-                    json!({"type":"layout","workspace":workspace.to_string(),"tab":(tab+1).to_string(),
-                    "focus":tab_ref.layout.focus.0.to_string(),"tree":tab_ref.layout.to_tree()}),
-                )
-            }
-            "layout.apply" => {
-                reject_api_fields(
-                    p,
-                    &[
-                        "workspace",
-                        "workspace_id",
-                        "tab",
-                        "tab_id",
-                        "focus",
-                        "tree",
-                    ],
-                )?;
-                let workspace = self.optional_socket_workspace(p)?.unwrap_or(self.active_ws);
-                let tab = self
-                    .optional_socket_tab(workspace, p, "tab", "tab_id")?
-                    .unwrap_or_else(|| {
-                        self.workspaces
-                            .get(workspace)
-                            .map(|ws| ws.active_tab)
-                            .unwrap_or(usize::MAX)
-                    });
-                let tree = crate::api::topology::parse_tree(p.get("tree").ok_or_else(|| {
-                    (
-                        "invalid_request".to_string(),
-                        "layout.apply needs a tree".to_string(),
-                    )
-                })?)?;
-                let tab_ref = self
-                    .workspaces
-                    .get_mut(workspace)
-                    .and_then(|ws| ws.tabs.get_mut(tab))
-                    .ok_or_else(not_found)?;
-                if !tab_ref.is_renameable() {
-                    return Err((
-                        "invalid_request".to_string(),
-                        "dashboard tabs do not have a mutable pane layout".to_string(),
-                    ));
-                }
-                let focus = match p.get("focus") {
-                    None => tab_ref.layout.focus,
-                    Some(value) => PaneId(parse_u32_value(value, "focus")?),
-                };
-                tab_ref
-                    .layout
-                    .apply_tree(&tree, focus)
-                    .map_err(|message| ("invalid_request".to_string(), message.to_string()))?;
-                self.session_dirty = true;
-                self.emit_event(
-                    "layout.applied",
-                    json!({"workspace":workspace.to_string(),"tab":(tab+1).to_string()}),
-                );
-                Ok(
-                    json!({"type":"layout_applied","workspace":workspace.to_string(),"tab":(tab+1).to_string()}),
-                )
-            }
-            "layout.set_split_ratio" => {
-                reject_api_fields(
-                    p,
-                    &[
-                        "workspace",
-                        "workspace_id",
-                        "tab",
-                        "tab_id",
-                        "path",
-                        "ratio",
-                    ],
-                )?;
-                let workspace = self.optional_socket_workspace(p)?.unwrap_or(self.active_ws);
-                let tab = self
-                    .optional_socket_tab(workspace, p, "tab", "tab_id")?
-                    .unwrap_or_else(|| {
-                        self.workspaces
-                            .get(workspace)
-                            .map(|ws| ws.active_tab)
-                            .unwrap_or(usize::MAX)
-                    });
-                let path = crate::api::topology::split_path(p)?;
-                let ratio = p.get("ratio").and_then(Value::as_f64).ok_or_else(|| {
-                    (
-                        "invalid_request".to_string(),
-                        "ratio must be a finite number".to_string(),
-                    )
-                })?;
-                if !ratio.is_finite() || !(0.0..=1.0).contains(&ratio) {
-                    return Err((
-                        "invalid_request".to_string(),
-                        "ratio must be between 0 and 1".to_string(),
-                    ));
-                }
-                let tab_ref = self
-                    .workspaces
-                    .get_mut(workspace)
-                    .and_then(|ws| ws.tabs.get_mut(tab))
-                    .ok_or_else(not_found)?;
-                if !tab_ref.layout.try_set_ratio(
-                    crate::api::topology::logical_area(),
-                    &path,
-                    ratio as f32,
-                ) {
-                    return Err((
-                        "not_found".to_string(),
-                        "path does not identify a split".to_string(),
-                    ));
-                }
-                self.session_dirty = true;
-                self.emit_event("layout.ratio_changed", json!({"workspace":workspace.to_string(),"tab":(tab+1).to_string(),"path":p["path"],"ratio":ratio}));
-                Ok(
-                    json!({"type":"layout_split_ratio","workspace":workspace.to_string(),"tab":(tab+1).to_string(),"ratio":ratio}),
-                )
-            }
+            "tab.rename" => self.api_tab_rename(method, p),
+            "tab.close" => self.api_tab_close(method, p),
+            "layout.export" => self.api_layout_export(method, p),
+            "layout.apply" => self.api_layout_apply(method, p),
+            "layout.set_split_ratio" => self.api_layout_set_split_ratio(method, p),
             // ── panes / agents ──
-            "pane.focus" => {
-                let id = self.resolve_pane(p)?.ok_or_else(not_found)?;
-                self.focus_pane_global(id);
-                Ok(json!({"type":"ok"}))
-            }
-            "pane.focus_direction" => {
-                reject_api_fields(p, &["pane", "direction"])?;
-                let pane = self.resolve_pane_or_focus(p)?;
-                let direction = crate::api::topology::direction(p)?;
-                let (workspace, tab) = self.pane_location(pane).ok_or_else(not_found)?;
-                let next = self.workspaces[workspace].tabs[tab]
-                    .layout
-                    .neighbor(crate::api::topology::logical_area(), pane, direction)
-                    .ok_or_else(|| {
-                        (
-                            "not_found".to_string(),
-                            "no pane exists in that direction".to_string(),
-                        )
-                    })?;
-                self.focus_pane_global(next);
-                self.emit_event("pane.focused", json!({"pane":next.0.to_string()}));
-                Ok(json!({"type":"pane_focus","pane":next.0.to_string()}))
-            }
-            "pane.resize" => {
-                reject_api_fields(p, &["pane", "direction", "cells"])?;
-                let pane = self.resolve_pane_or_focus(p)?;
-                let direction = crate::api::topology::direction(p)?;
-                let cells = p.get("cells").and_then(Value::as_i64).unwrap_or(1);
-                if !(1..=1000).contains(&cells) {
-                    return Err((
-                        "invalid_request".to_string(),
-                        "cells must be between 1 and 1000".to_string(),
-                    ));
-                }
-                let (workspace, tab) = self.pane_location(pane).ok_or_else(not_found)?;
-                let layout = &mut self.workspaces[workspace].tabs[tab].layout;
-                let previous = layout.focus;
-                layout.focus = pane;
-                let area = if self.last_pane_area.width > 1 && self.last_pane_area.height > 1 {
-                    self.last_pane_area
-                } else {
-                    crate::api::topology::logical_area()
-                };
-                let changed = layout.resize_focused(area, direction, cells as i16);
-                layout.focus = previous;
-                if !changed {
-                    return Err((
-                        "not_found".to_string(),
-                        "no matching divider can resize this pane".to_string(),
-                    ));
-                }
-                self.session_dirty = true;
-                self.emit_event(
-                    "pane.resized",
-                    json!({"pane":pane.0.to_string(),"cells":cells}),
-                );
-                Ok(json!({"type":"pane_resize","pane":pane.0.to_string()}))
-            }
-            "pane.zoom" => {
-                reject_api_fields(p, &["pane", "enabled"])?;
-                let pane = self.resolve_pane_or_focus(p)?;
-                let enabled = match p.get("enabled") {
-                    None => !self.zoomed,
-                    Some(Value::Bool(enabled)) => *enabled,
-                    Some(_) => {
-                        return Err((
-                            "invalid_request".to_string(),
-                            "enabled must be a boolean".to_string(),
-                        ))
-                    }
-                };
-                self.focus_pane_global(pane);
-                self.zoomed = enabled;
-                let pane = self.layout().focus;
-                self.emit_event(
-                    "pane.zoomed",
-                    json!({"pane":pane.0.to_string(),"enabled":enabled}),
-                );
-                Ok(json!({"type":"pane_zoom","pane":pane.0.to_string(),"enabled":enabled}))
-            }
-            "pane.rename" => {
-                reject_api_fields(p, &["pane", "name"])?;
-                let pane = self.resolve_pane(p)?.ok_or_else(not_found)?;
-                let name = p
-                    .get("name")
-                    .and_then(Value::as_str)
-                    .ok_or_else(|| {
-                        (
-                            "invalid_request".to_string(),
-                            "name must be a string".to_string(),
-                        )
-                    })?
-                    .trim();
-                if !name.is_empty() && !valid_agent_name(name) {
-                    return Err((
-                        "invalid_request".to_string(),
-                        "name must match [a-z][a-z0-9_-]{0,31}".to_string(),
-                    ));
-                }
-                self.set_agent_name(pane, (!name.is_empty()).then_some(name));
-                self.emit_event("pane.renamed", json!({"pane":pane.0.to_string(),"name":if name.is_empty(){Value::Null}else{json!(name)}}));
-                Ok(
-                    json!({"type":"pane_rename","pane":pane.0.to_string(),"name":if name.is_empty(){Value::Null}else{json!(name)}}),
-                )
-            }
-            "pane.swap" => {
-                reject_api_fields(p, &["pane", "with"])?;
-                let first = self.resolve_pane(p)?.ok_or_else(not_found)?;
-                let second = PaneId(parse_u32_value(
-                    p.get("with").ok_or_else(|| {
-                        (
-                            "invalid_request".to_string(),
-                            "with must be a pane id".to_string(),
-                        )
-                    })?,
-                    "with",
-                )?);
-                let first_location = self.pane_location(first).ok_or_else(not_found)?;
-                let second_location = self.pane_location(second).ok_or_else(not_found)?;
-                if first_location != second_location {
-                    return Err((
-                        "invalid_request".to_string(),
-                        "panes must belong to the same tab".to_string(),
-                    ));
-                }
-                let layout = &mut self.workspaces[first_location.0].tabs[first_location.1].layout;
-                if !layout.swap_panes(first, second) {
-                    return Err((
-                        "invalid_request".to_string(),
-                        "panes could not be swapped".to_string(),
-                    ));
-                }
-                self.session_dirty = true;
-                self.emit_event(
-                    "pane.swapped",
-                    json!({"pane":first.0.to_string(),"with":second.0.to_string()}),
-                );
-                Ok(
-                    json!({"type":"pane_swap","pane":first.0.to_string(),"with":second.0.to_string()}),
-                )
-            }
+            "pane.focus" => self.api_pane_focus(method, p),
+            "pane.focus_direction" => self.api_pane_focus_direction(method, p),
+            "pane.resize" => self.api_pane_resize(method, p),
+            "pane.zoom" => self.api_pane_zoom(method, p),
+            "pane.rename" => self.api_pane_rename(method, p),
+            "pane.swap" => self.api_pane_swap(method, p),
             // `attach.pane` (docs/18 WA-2): focus a pane and zoom it, so a client
             // attaching next opens straight into that fullscreen terminal.
-            "attach.pane" => {
-                let id = self.resolve_pane(p)?.ok_or_else(not_found)?;
-                self.focus_pane_global(id);
-                self.zoomed = true;
-                Ok(json!({"type":"ok","pane": id.0.to_string()}))
-            }
-            "agent.list" => {
-                let focus = self.layout().focus;
-                let mut arr = Vec::new();
-                for (wi, ws) in self.workspaces.iter().enumerate() {
-                    // Node-level context, identical for every pane in the node.
-                    // `project` deliberately repeats `workspace_name` so a consumer
-                    // can use one field name across `agent.list` *and*
-                    // `pane.agent_status_changed` without the label flip-flopping
-                    // between the node's label and its folder basename (docs/24).
-                    let branch = ws.branch.clone();
-                    let repo = ws
-                        .worktree
-                        .as_ref()
-                        .map(|m| m.common_dir.to_string_lossy().to_string());
-                    // Resolved when the membership was built (docs/18 WT) — this
-                    // runs on the app loop, so it must stay a field read.
-                    let is_worktree = ws.worktree.as_ref().is_some_and(|m| m.linked);
-                    for (ti, tab) in ws.tabs.iter().enumerate() {
-                        for id in tab.layout.leaves() {
-                            let Some(s) = self.status.get(&id) else {
-                                continue;
-                            };
-                            // Only real agent sessions, not the shells behind tabs.
-                            if !(self.manifests.is_agent(&s.agent)
-                                || s.agent_session.is_some()
-                                || s.agent_report.is_some())
-                            {
-                                continue;
-                            }
-                            let cwd = self
-                                .panes
-                                .get(&id)
-                                .map(|p| p.cwd.to_string_lossy().to_string())
-                                .unwrap_or_default();
-                            let terminal_id = self
-                                .panes
-                                .get(&id)
-                                .and_then(|pane| pane.terminal_runtime())
-                                .map(|runtime| runtime.terminal_id.clone());
-                            // The agent's own session id, when luvus knows it
-                            // exactly: reported by the integration hook, or set
-                            // because luvus launched it (resume/fork). `null`
-                            // means unbound — nothing is guessed here, so this
-                            // doubles as "is this pane's session actually known?"
-                            let session = s.agent_session.as_ref().map(|a| a.session_id.clone());
-                            arr.push(json!({
-                                "pane": id.0.to_string(), "agent": s.agent,
-                                "terminal_id": terminal_id,
-                                "name": self.agent_name_for(id),
-                                "status": state_str(s.state),
-                                "authority":s.identity_source,
-                                "state_source":s.state_source,
-                                "session": session,
-                                "workspace": wi.to_string(), "workspace_name": ws.name,
-                                // The one selector that survives reordering:
-                                // `workspace` is a positional index that shifts
-                                // when an earlier workspace closes, and
-                                // `workspace_name` is user-editable.
-                                "workspace_id": ws.id,
-                                "project": ws.name, "cwd": cwd,
-                                "branch": branch, "repo": repo, "worktree": is_worktree,
-                                "tab": (ti + 1).to_string(), "focused": id == focus,
-                            }));
-                        }
-                    }
-                }
-                Ok(json!({"type":"agent_list","agents":arr}))
-            }
+            "attach.pane" => self.api_attach_pane(method, p),
+            "agent.list" => self.api_agent_list(method, p),
             // Give a pane's agent a live alias (or clear it) so `agent.send` /
             // `agent.keys` / `agent.read` can address it by name. Ephemeral.
-            "agent.name" => {
-                let pane = self.resolve_pane(p)?.ok_or_else(not_found)?;
-                if p.get("clear").and_then(|v| v.as_bool()).unwrap_or(false) {
-                    self.set_agent_name(pane, None);
-                    return Ok(
-                        json!({"type":"agent_name","pane": pane.0.to_string(), "name": Value::Null}),
-                    );
-                }
-                let name = p.get("name").and_then(|v| v.as_str()).unwrap_or("");
-                if !valid_agent_name(name) {
-                    return Err((
-                        "invalid_request".to_string(),
-                        "name must match [a-z][a-z0-9_-]{0,31}".to_string(),
-                    ));
-                }
-                self.set_agent_name(pane, Some(name));
-                Ok(json!({"type":"agent_name","pane": pane.0.to_string(), "name": name}))
-            }
+            "agent.name" => self.api_agent_name(method, p),
             // Fork a live agent's native session into a sibling pane. Target
             // resolution matches agent.send/get: alias, pane id, or unique kind.
-            "agent.fork" => {
-                let pane = self.resolve_agent_target(p)?;
-                let focus = match p.get("focus") {
-                    None => true,
-                    Some(Value::Bool(v)) => *v,
-                    Some(_) => {
-                        return Err((
-                            "invalid_request".to_string(),
-                            "focus must be a boolean".to_string(),
-                        ))
-                    }
-                };
-                let name = match p.get("name") {
-                    None => None,
-                    Some(Value::String(v)) if valid_agent_name(v) => Some(v.as_str()),
-                    Some(_) => {
-                        return Err((
-                            "invalid_request".to_string(),
-                            "name must match [a-z][a-z0-9_-]{0,31}".to_string(),
-                        ))
-                    }
-                };
-                let forked = self
-                    .fork_agent_pane(pane, focus)
-                    .map_err(agent_fork_error)?;
-                if let Some(alias) = name {
-                    self.set_agent_name(forked.pane, Some(alias));
-                }
-                Ok(json!({
-                    "type": "agent_fork",
-                    "from": forked.from.0.to_string(),
-                    "pane": forked.pane.0.to_string(),
-                    "agent": forked.agent,
-                    "name": name,
-                    "workspace": forked.workspace.to_string(),
-                    "tab": (forked.tab + 1).to_string(),
-                    "focused": focus,
-                }))
-            }
+            "agent.fork" => self.api_agent_fork(method, p),
             // Submit a prompt to a target agent: paste the text (bracketed when the
             // child asked for it), then send Enter once the paste has landed.
-            "agent.send" => {
-                let id = self.resolve_agent_target(p)?;
-                if !self.is_agent_pane(id) {
-                    return Err((
-                        "agent_not_ready".to_string(),
-                        "target pane is not a running agent".to_string(),
-                    ));
-                }
-                let text = p.get("text").and_then(|v| v.as_str()).unwrap_or("");
-                if text.is_empty() {
-                    return Err((
-                        "invalid_request".to_string(),
-                        "agent send text must not be empty".to_string(),
-                    ));
-                }
-                let pane = self.panes.get(&id).ok_or_else(|| {
-                    (
-                        "send_failed".to_string(),
-                        "target pane closed before input was queued".to_string(),
-                    )
-                })?;
-                pane.try_submit_text_with_settle(text, AGENT_MESSAGE_SETTLE)
-                    .map_err(|message| ("send_failed".to_string(), message))?;
-                let (agent, status) = self
-                    .status
-                    .get(&id)
-                    .map(|s| (s.agent.clone(), state_str(s.state).to_string()))
-                    .unwrap_or_default();
-                Ok(json!({"type":"agent_send","pane": id.0.to_string(),
-                          "agent": agent, "status": status, "name": self.agent_name_for(id)}))
-            }
+            "agent.send" => self.api_agent_send(method, p),
             // Send named control keys (enter, esc, ctrl+c, up, …) to a target agent,
             // e.g. to answer a blocked approval prompt. All keys validate first.
-            "agent.keys" => {
-                reject_api_fields(p, &["target", "keys"])?;
-                let id = self.resolve_agent_target(p)?;
-                if !self.is_agent_pane(id) {
-                    return Err((
-                        "agent_not_ready".to_string(),
-                        "target pane is not a running agent".to_string(),
-                    ));
-                }
-                let keys = p.get("keys").and_then(|v| v.as_array()).ok_or_else(|| {
-                    (
-                        "invalid_request".to_string(),
-                        "agent keys must be a non-empty array".to_string(),
-                    )
-                })?;
-                if keys.is_empty() {
-                    return Err((
-                        "invalid_request".to_string(),
-                        "agent keys must be a non-empty array".to_string(),
-                    ));
-                }
-                let mut bytes = Vec::new();
-                for key in keys {
-                    let key = key.as_str().ok_or_else(|| {
-                        (
-                            "invalid_request".to_string(),
-                            "every agent key must be a string".to_string(),
-                        )
-                    })?;
-                    bytes.extend(key_to_bytes(key).ok_or_else(|| {
-                        ("invalid_request".to_string(), format!("unknown key: {key}"))
-                    })?);
-                }
-                self.panes
-                    .get(&id)
-                    .ok_or_else(not_found)?
-                    .try_send(&bytes)
-                    .map_err(|message| ("send_failed".to_string(), message))?;
-                Ok(json!({"type":"ok","pane": id.0.to_string()}))
-            }
+            "agent.keys" => self.api_agent_keys(method, p),
             // Read a target agent's output, addressed by name or pane id.
-            "agent.read" => {
-                let id = self.resolve_agent_target(p)?;
-                let lines = p.get("lines").and_then(|v| v.as_u64()).unwrap_or(200) as u16;
-                // `visible` = the current screen; anything else = recent output
-                // (soft wraps joined), the default and best for transcripts.
-                let source = p.get("source").and_then(|v| v.as_str()).unwrap_or("recent");
-                let text = self
-                    .panes
-                    .get(&id)
-                    .and_then(|pane| {
-                        pane.engine.lock().ok().map(|e| {
-                            if source == "visible" {
-                                e.visible_rows().join("\n")
-                            } else {
-                                e.detection_text(lines)
-                            }
-                        })
-                    })
-                    .unwrap_or_default();
-                Ok(json!({"type":"agent_read","pane": id.0.to_string(), "text": text}))
-            }
+            "agent.read" => self.api_agent_read(method, p),
             // One agent's live info, resolved by name / pane id / kind — what to
             // check before deciding how to answer a blocked agent.
-            "agent.get" => {
-                let id = self.resolve_agent_target(p)?;
-                let s = self.status.get(&id);
-                let cwd = self
-                    .panes
-                    .get(&id)
-                    .map(|pn| pn.cwd.display().to_string())
-                    .unwrap_or_default();
-                let (agent, status, authority, state_source) = s
-                    .map(|s| {
-                        (
-                            s.agent.clone(),
-                            state_str(s.state).to_string(),
-                            s.identity_source,
-                            s.state_source,
-                        )
-                    })
-                    .unwrap_or_default();
-                let session =
-                    s.and_then(|s| s.agent_session.as_ref().map(|a| a.session_id.clone()));
-                Ok(json!({"type":"agent","pane": id.0.to_string(),
-                          "name": self.agent_name_for(id), "agent": agent,
-                          "status": status, "authority":authority,
-                          "state_source":state_source, "session": session, "cwd": cwd}))
-            }
-            "agent.explain" => {
-                reject_api_fields(p, &["target", "pane"])?;
-                if p.get("target").is_some() == p.get("pane").is_some() {
-                    return Err((
-                        "invalid_request".to_string(),
-                        "agent.explain needs exactly one of target or pane".to_string(),
-                    ));
-                }
-                if let Some(target) = p.get("target") {
-                    let valid = target
-                        .as_str()
-                        .is_some_and(|target| !target.is_empty() && target.chars().count() <= 128);
-                    if !valid {
-                        return Err((
-                            "invalid_request".to_string(),
-                            "target must be a non-empty string of at most 128 characters"
-                                .to_string(),
-                        ));
-                    }
-                }
-                let id = if p.get("target").is_some() {
-                    self.resolve_agent_pane(p).ok_or_else(agent_not_found)?
-                } else {
-                    self.resolve_pane(p)?.ok_or_else(not_found)?
-                };
-                Ok(self.agent_explanation(id))
-            }
-            "agent.report" => {
-                reject_api_fields(
-                    p,
-                    &[
-                        "pane",
-                        "source",
-                        "agent",
-                        "status",
-                        "message",
-                        "session_id",
-                        "sequence",
-                        "ttl_s",
-                    ],
-                )?;
-                // `target` is not an accepted field here, so `pane` is the only
-                // explicit target and a miss is terminal — the same rule
-                // `agent.explain` applies, with no fallback to the focused pane.
-                let id = self.resolve_pane(p)?.ok_or_else(not_found)?;
-                let source = required_report_source(p)?;
-                let agent = p.get("agent").and_then(Value::as_str).ok_or_else(|| {
-                    (
-                        "invalid_request".to_string(),
-                        "agent.report needs an agent".to_string(),
-                    )
-                })?;
-                if !valid_agent_name(agent) {
-                    return Err((
-                        "invalid_request".to_string(),
-                        "agent must match [a-z][a-z0-9_-]{0,31}".to_string(),
-                    ));
-                }
-                let state = p
-                    .get("status")
-                    .and_then(Value::as_str)
-                    .and_then(parse_agent_wait_state)
-                    .ok_or_else(|| {
-                        (
-                            "invalid_request".to_string(),
-                            "status must be idle, working, blocked, or done".to_string(),
-                        )
-                    })?;
-                let message =
-                    optional_bounded_string(p, "message", MAX_AGENT_REPORT_MESSAGE_CHARS)?;
-                let session_id = optional_bounded_string(p, "session_id", 512)?;
-                let ttl_s = p.get("ttl_s").and_then(Value::as_u64).unwrap_or(3600);
-                if !(1..=MAX_AGENT_REPORT_TTL_S).contains(&ttl_s) {
-                    return Err((
-                        "invalid_request".to_string(),
-                        "ttl_s must be between 1 and 86400".to_string(),
-                    ));
-                }
-                let now = Instant::now();
-                let current = self
-                    .status
-                    .get(&id)
-                    .and_then(|status| status.agent_report.as_ref());
-                if current.is_some_and(|report| report.source != source) {
-                    return Err((
-                        "authority_conflict".to_string(),
-                        "another integration owns this pane; release it first".to_string(),
-                    ));
-                }
-                let sequence = match p.get("sequence") {
-                    Some(Value::Number(number)) => number.as_u64().ok_or_else(|| {
-                        (
-                            "invalid_request".to_string(),
-                            "sequence must be a non-negative integer".to_string(),
-                        )
-                    })?,
-                    Some(_) => {
-                        return Err((
-                            "invalid_request".to_string(),
-                            "sequence must be a non-negative integer".to_string(),
-                        ))
-                    }
-                    None => current.map_or(1, |report| report.sequence.saturating_add(1)),
-                };
-                if current.is_some_and(|report| sequence <= report.sequence) {
-                    return Err((
-                        "stale_report".to_string(),
-                        "sequence must increase for this authority".to_string(),
-                    ));
-                }
-                let (changed, cwd, project, branch) = {
-                    let status = self.status.get_mut(&id).ok_or_else(not_found)?;
-                    let changed = status.state != state || status.agent != agent;
-                    status.agent = agent.to_string();
-                    status.state = state;
-                    status.candidate = state;
-                    status.candidate_since = now;
-                    status.prev_working = state == State::Working;
-                    status.done = state == State::Done;
-                    status.identity_source = "integration_report";
-                    status.state_source = "integration_report";
-                    status.rule_priority = None;
-                    status.rule_region = None;
-                    status.blocked_hint =
-                        (state == State::Blocked).then(|| message.clone()).flatten();
-                    status.agent_report = Some(AgentReport {
-                        source: source.clone(),
-                        agent: agent.to_string(),
-                        state,
-                        message: message.clone(),
-                        sequence,
-                        expires_at: now + Duration::from_secs(ttl_s),
-                    });
-                    if let Some(session_id) = session_id.as_ref() {
-                        status.agent_session = Some(AgentSession {
-                            agent: agent.to_string(),
-                            session_id: session_id.clone(),
-                        });
-                    }
-                    let cwd = self
-                        .panes
-                        .get(&id)
-                        .map(|pane| pane.cwd.display().to_string())
-                        .unwrap_or_default();
-                    let (project, branch) = self
-                        .workspace_of_pane(id)
-                        .map(|workspace| (workspace.name.clone(), workspace.branch.clone()))
-                        .unwrap_or_default();
-                    (changed, cwd, project, branch)
-                };
-                self.emit_event(
-                    "agent.authority_reported",
-                    json!({"pane":id.0.to_string(), "source":source, "agent":agent, "status":state_str(state), "sequence":sequence, "ttl_s":ttl_s}),
-                );
-                log_agent_authority(id, agent, crate::logging::Outcome::Ok);
-                self.reconcile_durable_active_targets(Some(id));
-                if changed {
-                    self.emit_event(
-                        "pane.agent_status_changed",
-                        json!({"pane":id.0.to_string(), "status":state_str(state), "agent":agent, "cwd":cwd, "project":project, "branch":branch, "authority":"integration_report"}),
-                    );
-                    self.wake_active_agent_automations(id);
-                }
-                self.check_agent_waits(id);
-                Ok(json!({
-                    "type":"agent_report", "pane":id.0.to_string(),
-                    "agent":agent, "status":state_str(state), "source":source,
-                    "sequence":sequence, "ttl_s":ttl_s,
-                }))
-            }
-            "agent.release" => {
-                reject_api_fields(p, &["pane", "source"])?;
-                // `target` is not an accepted field here, so `pane` is the only
-                // explicit target and a miss is terminal — the same rule
-                // `agent.explain` applies, with no fallback to the focused pane.
-                let id = self.resolve_pane(p)?.ok_or_else(not_found)?;
-                let source = required_report_source(p)?;
-                let status = self.status.get_mut(&id).ok_or_else(not_found)?;
-                let Some(report) = status.agent_report.as_ref() else {
-                    return Err((
-                        "not_found".to_string(),
-                        "pane has no integration authority".to_string(),
-                    ));
-                };
-                if report.source != source {
-                    return Err((
-                        "authority_conflict".to_string(),
-                        "source does not own this pane".to_string(),
-                    ));
-                }
-                status.agent_report = None;
-                status.force_detect = true;
-                let agent = status.agent.clone();
-                self.emit_event(
-                    "agent.authority_released",
-                    json!({"pane":id.0.to_string(), "source":source, "reason":"released"}),
-                );
-                log_agent_authority(id, &agent, crate::logging::Outcome::Ok);
-                Ok(json!({"type":"agent_release", "pane":id.0.to_string()}))
-            }
-            "agent.wait" => Err((
-                "internal".to_string(),
-                "agent.wait must be dispatched through the event-driven waiter".to_string(),
-            )),
+            "agent.get" => self.api_agent_get(method, p),
+            "agent.explain" => self.api_agent_explain(method, p),
+            "agent.report" => self.api_agent_report(method, p),
+            "agent.release" => self.api_agent_release(method, p),
+            "agent.wait" => self.api_agent_wait(method, p),
             // Resumable sessions discovered on disk (the AGENTS sidebar list).
-            "agent.sessions" => {
-                self.refresh_resumable();
-                let arr: Vec<Value> = self
-                    .resumable
-                    .iter()
-                    .map(|s| {
-                        json!({
-                            "agent": s.agent,
-                            "session_id": s.session_id,
-                            "cwd": s.cwd.display().to_string(),
-                        })
-                    })
-                    .collect();
-                Ok(json!({"type":"session_list","sessions":arr}))
-            }
-            "agent.resume" => {
-                self.refresh_resumable();
-                let sid = p.get("session_id").and_then(|v| v.as_str()).unwrap_or("");
-                let idx = self.resumable.iter().position(|s| s.session_id == sid);
-                match idx {
-                    Some(i) => {
-                        self.resume_session(i);
-                        Ok(json!({"type":"ok"}))
-                    }
-                    None => Err((
-                        "not_found".to_string(),
-                        "no resumable session with that id".to_string(),
-                    )),
-                }
-            }
+            "agent.sessions" => self.api_agent_sessions(method, p),
+            "agent.resume" => self.api_agent_resume(method, p),
             // ── ui / appearance ──
-            "ui.sidebar" => {
-                // `side` selects left (default) or right (docs/29).
-                let side = match p.get("side").and_then(|v| v.as_str()) {
-                    Some("right") => crate::app::Side::Right,
-                    _ => crate::app::Side::Left,
-                };
-                if let Some(w) = param_usize(p, "width") {
-                    self.set_side_width(side, w as u16);
-                }
-                if let Some(v) = p.get("visible").and_then(|v| v.as_bool()) {
-                    self.sidebars.get_mut(side).visible = v;
-                }
-                let s = self.sidebars.get(side);
-                Ok(json!({
-                    "type": "ok",
-                    "width": s.width,
-                    "visible": s.visible,
-                }))
-            }
+            "ui.sidebar" => self.api_ui_sidebar(method, p),
             // A module pushes rows into its sidebar dock (docs/29, DOCK-4).
             // A one-line confirmation, the same transient toast a copy shows.
-            "ui.toast" => {
-                let text = req_str(p, "text")?;
-                self.show_toast(text.chars().take(120).collect::<String>());
-                Ok(json!({"type":"ok"}))
-            }
-            "ui.dock.push" => {
-                let id = p.get("id").and_then(|v| v.as_str()).unwrap_or("");
-                if id.is_empty() {
-                    return Ok(json!({"type":"error","message":"dock id required"}));
-                }
-                let title = p
-                    .get("title")
-                    .and_then(|v| v.as_str())
-                    .map(|s| s.to_string());
-                let placement = match p.get("placement").and_then(|v| v.as_str()) {
-                    Some("right") | Some("sidebar.right") => crate::app::Side::Right,
-                    _ => crate::app::Side::Left,
-                };
-                let rows = p
-                    .get("rows")
-                    .and_then(|v| v.as_array())
-                    .map(|arr| {
-                        arr.iter()
-                            .map(|r| {
-                                // A span with no text draws nothing, so drop it
-                                // here: a malformed list (`["a", "b"]`, `[{}]`)
-                                // then parses to no spans and the row falls
-                                // back to `text` instead of rendering blank.
-                                let spans: Vec<crate::app::DockSpan> = r
-                                    .get("spans")
-                                    .and_then(|v| v.as_array())
-                                    .map(|items| {
-                                        items
-                                            .iter()
-                                            .filter_map(|sp| {
-                                                let text = sp.get("text")?.as_str()?;
-                                                (!text.is_empty()).then(|| crate::app::DockSpan {
-                                                    text: text.to_string(),
-                                                    tone: sp
-                                                        .get("tone")
-                                                        .and_then(|v| v.as_str())
-                                                        .map(|s| s.to_string()),
-                                                })
-                                            })
-                                            .collect()
-                                    })
-                                    .unwrap_or_default();
-                                // `text` is what a click hands to the action. A
-                                // spans-only row gets the joined span text, so
-                                // moving a row from `text` to `spans` cannot
-                                // leave its action with an empty target.
-                                let text = r
-                                    .get("text")
-                                    .and_then(|v| v.as_str())
-                                    .map(|s| s.to_string())
-                                    .unwrap_or_else(|| {
-                                        spans.iter().map(|sp| sp.text.as_str()).collect()
-                                    });
-                                crate::app::DockRow {
-                                    text,
-                                    dot: r
-                                        .get("dot")
-                                        .and_then(|v| v.as_str())
-                                        .map(|s| s.to_string()),
-                                    tone: r
-                                        .get("tone")
-                                        .and_then(|v| v.as_str())
-                                        .map(|s| s.to_string()),
-                                    spans,
-                                    action: r
-                                        .get("action")
-                                        .and_then(|v| v.as_str())
-                                        .map(|s| s.to_string()),
-                                    value: r
-                                        .get("value")
-                                        .and_then(|v| v.as_str())
-                                        .map(|s| s.to_string()),
-                                    // Right-click menu for this row (docs/52).
-                                    // Absent — every module written before this —
-                                    // leaves the row with no menu, as before. An
-                                    // entry with no `action` is a divider.
-                                    menu: r
-                                        .get("menu")
-                                        .and_then(|v| v.as_array())
-                                        .map(|items| {
-                                            items
-                                                .iter()
-                                                .map(|it| crate::app::DockRowMenuItem {
-                                                    title: it
-                                                        .get("title")
-                                                        .and_then(|v| v.as_str())
-                                                        .unwrap_or("")
-                                                        .to_string(),
-                                                    action: it
-                                                        .get("action")
-                                                        .and_then(|v| v.as_str())
-                                                        .unwrap_or("")
-                                                        .to_string(),
-                                                    value: it
-                                                        .get("value")
-                                                        .and_then(|v| v.as_str())
-                                                        .map(|s| s.to_string()),
-                                                    destructive: it
-                                                        .get("destructive")
-                                                        .and_then(|v| v.as_bool())
-                                                        .unwrap_or(false),
-                                                })
-                                                .collect()
-                                        })
-                                        .unwrap_or_default(),
-                                }
-                            })
-                            .collect()
-                    })
-                    .unwrap_or_default();
-                self.push_module_dock(id, title, placement, rows);
-                Ok(json!({"type":"ok"}))
-            }
-            "ui.dock.list" => {
-                let arr: Vec<Value> = self
-                    .docks_flat()
-                    .iter()
-                    .map(|k| {
-                        let side = match self.sidebars.side_of(k) {
-                            Some(crate::app::Side::Right) => "right",
-                            _ => "left",
-                        };
-                        json!({"id": k.id(), "side": side})
-                    })
-                    .collect();
-                Ok(json!({"type":"dock_list","docks":arr}))
-            }
-            "ui.dock.move" => {
-                let id = p.get("id").and_then(|v| v.as_str()).unwrap_or("");
-                if id.is_empty() {
-                    return Ok(json!({"type":"error","message":"dock id required"}));
-                }
-                let side = match p.get("side").and_then(|v| v.as_str()) {
-                    Some("right") => crate::app::Side::Right,
-                    _ => crate::app::Side::Left,
-                };
-                if self.move_dock(&crate::app::DockKind::from_id(id), side) {
-                    Ok(json!({"type":"ok"}))
-                } else {
-                    Ok(json!({"type":"error","message":"sidebar is full (max 3 docks)"}))
-                }
-            }
-            "ui.bar.list" => {
-                let widgets: Vec<Value> = self
-                    .bar
-                    .declarations
-                    .iter()
-                    .map(|(key, declaration)| {
-                        let live = self.bar.widgets.get(key);
-                        let region = self
-                            .config
-                            .bars
-                            .region_for(key, declaration.region)
-                            .map(crate::bar::BarRegion::as_str);
-                        json!({
-                            "id": declaration.key.id,
-                            "owner": declaration.key.owner,
-                            "key": key,
-                            "title": declaration.title,
-                            "region": region,
-                            "default_region": declaration.region.as_str(),
-                            "priority": live.map_or(declaration.priority, |widget| widget.priority),
-                            "live": live.is_some(),
-                            "content": live.map(|widget| &widget.content),
-                            "compact_content": live.map(|widget| &widget.compact_content),
-                        })
-                    })
-                    .collect();
-                Ok(json!({"type":"bar_list","widgets":widgets}))
-            }
-            "ui.bar.push" => {
-                let id = req_str(p, "id")?;
-                let owner = p.get("owner").and_then(Value::as_str);
-                let declaration = self
-                    .bar
-                    .resolve_declaration(owner, id)
-                    .map_err(module_err)?
-                    .clone();
-                if declaration.key.owner == "core" {
-                    return Err(module_err("core bar widgets cannot be updated".into()));
-                }
-                let content: Vec<crate::bar::BarSegment> = serde_json::from_value(
-                    p.get("content")
-                        .cloned()
-                        .ok_or_else(|| ("invalid_request".into(), "content is required".into()))?,
-                )
-                .map_err(|error| {
-                    (
-                        "invalid_request".into(),
-                        format!("invalid content: {error}"),
-                    )
-                })?;
-                let compact: Vec<crate::bar::BarSegment> = match p.get("compact_content") {
-                    Some(value) => serde_json::from_value(value.clone()).map_err(|error| {
-                        (
-                            "invalid_request".into(),
-                            format!("invalid compact_content: {error}"),
-                        )
-                    })?,
-                    None => Vec::new(),
-                };
-                validate_bar_actions(self, &declaration.key.owner, &content)?;
-                validate_bar_actions(self, &declaration.key.owner, &compact)?;
-                let region = match p.get("region").and_then(Value::as_str) {
-                    Some("top-right" | "top") => crate::bar::BarRegion::TopRight,
-                    Some("bottom-right" | "bottom") => crate::bar::BarRegion::BottomRight,
-                    Some(other) => {
-                        return Err((
-                            "invalid_request".into(),
-                            format!("unknown bar region {other}"),
-                        ))
-                    }
-                    None => declaration.region,
-                };
-                let priority = match p.get("priority") {
-                    None => declaration.priority,
-                    Some(value) => value
-                        .as_u64()
-                        .filter(|value| *value <= u8::MAX as u64)
-                        .map(|value| value as u8)
-                        .ok_or_else(|| {
-                            (
-                                "invalid_request".into(),
-                                "priority must be an integer from 0 to 255".into(),
-                            )
-                        })?,
-                };
-                let widget = crate::bar::BarWidget::new(
-                    declaration.key.clone(),
-                    region,
-                    content,
-                    compact,
-                    priority,
-                )
-                .map_err(|error| ("invalid_request".into(), error))?;
-                self.bar
-                    .allow_push(&declaration.key.owner, Instant::now())
-                    .map_err(|error| ("rate_limited".into(), error))?;
-                let changed = self
-                    .bar
-                    .push_widget(widget)
-                    .map_err(|error| ("limit_exceeded".into(), error))?;
-                Ok(json!({"type":"ok","changed":changed,"key":declaration.key.canonical()}))
-            }
-            "ui.bar.move" => {
-                let declaration = self
-                    .bar
-                    .resolve_declaration(p.get("owner").and_then(Value::as_str), req_str(p, "id")?)
-                    .map_err(module_err)?
-                    .clone();
-                let region = match req_str(p, "region")? {
-                    "top-right" | "top" => Some(crate::bar::BarRegion::TopRight),
-                    "bottom-right" | "bottom" => Some(crate::bar::BarRegion::BottomRight),
-                    "off" => None,
-                    other => {
-                        return Err((
-                            "invalid_request".into(),
-                            format!("unknown bar region {other}"),
-                        ))
-                    }
-                };
-                let key = declaration.key.canonical();
-                if !self.config.bars.is_explicitly_placed(&key, region) {
-                    self.config.bars.place(&key, region);
-                    self.persist_config();
-                    self.bar.clear_geometry();
-                }
-                Ok(
-                    json!({"type":"ok","key":key,"region":region.map(crate::bar::BarRegion::as_str)}),
-                )
-            }
-            "ui.bar.remove" => {
-                let declaration = self
-                    .bar
-                    .resolve_declaration(p.get("owner").and_then(Value::as_str), req_str(p, "id")?)
-                    .map_err(module_err)?
-                    .clone();
-                if declaration.key.owner == "core" {
-                    return Err(module_err("core bar widgets cannot be removed".into()));
-                }
-                let removed = self.bar.remove_widget(&declaration.key.canonical());
-                Ok(json!({"type":"ok","removed":removed}))
-            }
-            "ui.notification.push" => {
-                let owner = p.get("owner").and_then(Value::as_str).map(String::from);
-                let text = req_str(p, "text")?.to_string();
-                let level: crate::bar::NotificationLevel = serde_json::from_value(
-                    p.get("level").cloned().unwrap_or_else(|| json!("info")),
-                )
-                .map_err(|error| ("invalid_request".into(), format!("invalid level: {error}")))?;
-                let action = opt_str(p, "action");
-                if let Some(owner) = owner.as_deref() {
-                    validate_bar_action(self, owner, action.as_deref())?;
-                } else if action.is_some() {
-                    return Err((
-                        "invalid_request".into(),
-                        "an actionable notification requires its module owner".into(),
-                    ));
-                }
-                let ttl_ms = match p.get("ttl_ms") {
-                    None => 4_000,
-                    Some(value) => value.as_u64().filter(|ttl| *ttl > 0).ok_or_else(|| {
-                        (
-                            "invalid_request".into(),
-                            "ttl_ms must be a positive integer".into(),
-                        )
-                    })?,
-                };
-                let notification = crate::bar::NotificationPush {
-                    owner,
-                    text,
-                    level,
-                    ttl_ms,
-                    action,
-                    value: opt_str(p, "value"),
-                    dedupe_key: opt_str(p, "dedupe_key"),
-                };
-                notification
-                    .validate()
-                    .map_err(|error| ("invalid_request".into(), error))?;
-                self.bar
-                    .allow_push(
-                        notification
-                            .owner
-                            .as_deref()
-                            .unwrap_or(crate::bar::UNOWNED_NOTIFICATION_OWNER),
-                        Instant::now(),
-                    )
-                    .map_err(|error| ("rate_limited".into(), error))?;
-                self.bar
-                    .push_notification(notification, Instant::now())
-                    .map_err(|error| ("invalid_request".into(), error))?;
-                Ok(json!({"type":"ok"}))
-            }
-            "ui.notification.clear" => {
-                let owner = p.get("owner").and_then(Value::as_str);
-                let removed = self
-                    .bar
-                    .clear_notifications(owner, p.get("dedupe_key").and_then(Value::as_str));
-                Ok(json!({"type":"ok","removed":removed}))
-            }
+            "ui.toast" => self.api_ui_toast(method, p),
+            "ui.dock.push" => self.api_ui_dock_push(method, p),
+            "ui.dock.list" => self.api_ui_dock_list(method, p),
+            "ui.dock.move" => self.api_ui_dock_move(method, p),
+            "ui.bar.list" => self.api_ui_bar_list(method, p),
+            "ui.bar.push" => self.api_ui_bar_push(method, p),
+            "ui.bar.move" => self.api_ui_bar_move(method, p),
+            "ui.bar.remove" => self.api_ui_bar_remove(method, p),
+            "ui.notification.push" => self.api_ui_notification_push(method, p),
+            "ui.notification.clear" => self.api_ui_notification_clear(method, p),
             // ── modules (docs/13) ──
-            "module.list" => {
-                let arr: Vec<Value> = self.modules.modules.iter().map(module_json).collect();
-                Ok(json!({"type":"module_list","modules":arr}))
-            }
-            "module.info" => {
-                let id = req_str(p, "id")?;
-                let m = self
-                    .modules
-                    .find(id)
-                    .ok_or_else(|| module_err(format!("no module {id}")))?;
-                Ok(json!({
-                    "type": "module_info",
-                    "id": m.id,
-                    "name": m.manifest.name,
-                    "version": m.manifest.version,
-                    "description": m.manifest.description,
-                    "enabled": m.enabled,
-                    "runnable": m.is_runnable(),
-                    "source": m.source,
-                    "root": m.root.display().to_string(),
-                    "warning": m.warning,
-                    "platforms": m.manifest.platforms,
-                    "actions": m.manifest.actions.iter()
-                        .map(|a| json!({"id": a.id, "title": a.title, "contexts": a.contexts})).collect::<Vec<_>>(),
-                    "panes": m.manifest.panes.iter()
-                        .map(|pe| json!({"id": pe.id, "title": pe.title, "placement": pe.placement})).collect::<Vec<_>>(),
-                    "bars": m.manifest.bars.iter()
-                        .map(|bar| json!({"id": bar.id, "title": bar.title, "region": bar.region.as_str(), "priority": bar.priority})).collect::<Vec<_>>(),
-                    "events": m.manifest.events.iter().map(|e| e.on.clone()).collect::<Vec<_>>(),
-                    "build_steps": m.manifest.build.len(),
-                }))
-            }
-            "module.link" => {
-                let path = req_str(p, "path")?;
-                let enabled = !p.get("disabled").and_then(|v| v.as_bool()).unwrap_or(false);
-                let source = p.get("source").and_then(|v| v.as_str()).map(String::from);
-                let id = self
-                    .module_link_with(std::path::Path::new(path), enabled, source)
-                    .map_err(module_err)?;
-                Ok(json!({"type":"module","id": id}))
-            }
-            "module.unlink" => {
-                self.module_unlink(req_str(p, "id")?).map_err(module_err)?;
-                Ok(json!({"type":"ok"}))
-            }
-            "module.uninstall" => {
-                self.module_uninstall(req_str(p, "id")?)
-                    .map_err(module_err)?;
-                Ok(json!({"type":"ok"}))
-            }
-            "module.enable" => {
-                self.module_set_enabled(req_str(p, "id")?, true)
-                    .map_err(module_err)?;
-                Ok(json!({"type":"ok"}))
-            }
-            "module.disable" => {
-                self.module_set_enabled(req_str(p, "id")?, false)
-                    .map_err(module_err)?;
-                Ok(json!({"type":"ok"}))
-            }
-            "module.action.list" => {
-                let mut arr = Vec::new();
-                for m in &self.modules.modules {
-                    for a in &m.manifest.actions {
-                        arr.push(json!({
-                            "module": m.id, "action": a.id,
-                            "qualified": format!("{}.{}", m.id, a.id),
-                            "title": a.title, "contexts": a.contexts,
-                            "runnable": m.is_runnable(),
-                        }));
-                    }
-                }
-                Ok(json!({"type":"module_action_list","actions":arr}))
-            }
-            "module.action.invoke" => {
-                let action = p
-                    .get("id")
-                    .or_else(|| p.get("action"))
-                    .and_then(|v| v.as_str())
-                    .filter(|s| !s.is_empty())
-                    .ok_or_else(|| {
-                        (
-                            "invalid_request".to_string(),
-                            "action id is required".to_string(),
-                        )
-                    })?;
-                let module = p.get("module").and_then(|v| v.as_str());
-                let log_id = self
-                    .module_invoke_action(action, module, "api")
-                    .map_err(module_err)?;
-                Ok(json!({"type":"module_command","log_id": log_id}))
-            }
-            "module.log.list" => {
-                let filter = p
-                    .get("id")
-                    .or_else(|| p.get("module"))
-                    .and_then(|v| v.as_str());
-                let limit = param_usize(p, "limit").unwrap_or(50);
-                let logs: Vec<Value> = self
-                    .module_logs
-                    .iter()
-                    .rev()
-                    .filter(|l| filter.is_none_or(|f| l.module_id == f))
-                    .take(limit)
-                    .map(|l| serde_json::to_value(l).unwrap_or(Value::Null))
-                    .collect();
-                Ok(json!({"type":"module_log_list","logs":logs}))
-            }
-            "module.config_dir" => {
-                let dir = self
-                    .module_config_dir(req_str(p, "id")?)
-                    .map_err(module_err)?;
-                Ok(json!({"type":"module_config_dir","dir": dir.display().to_string()}))
-            }
-            "module.pane.open" => {
-                let module = p
-                    .get("module")
-                    .or_else(|| p.get("id"))
-                    .and_then(|v| v.as_str())
-                    .filter(|s| !s.is_empty())
-                    .ok_or_else(|| {
-                        (
-                            "invalid_request".to_string(),
-                            "module id is required".to_string(),
-                        )
-                    })?;
-                let entrypoint = req_str(p, "entrypoint")?;
-                let placement = p.get("placement").and_then(|v| v.as_str());
-                let id = self
-                    .module_open_pane(module, entrypoint, placement, "api")
-                    .map_err(module_err)?;
-                Ok(json!({"type":"pane","pane": id.0.to_string()}))
-            }
+            "module.list" => self.api_module_list(method, p),
+            "module.info" => self.api_module_info(method, p),
+            "module.link" => self.api_module_link(method, p),
+            "module.unlink" => self.api_module_unlink(method, p),
+            "module.uninstall" => self.api_module_uninstall(method, p),
+            "module.enable" => self.api_module_enable(method, p),
+            "module.disable" => self.api_module_disable(method, p),
+            "module.action.list" => self.api_module_action_list(method, p),
+            "module.action.invoke" => self.api_module_action_invoke(method, p),
+            "module.log.list" => self.api_module_log_list(method, p),
+            "module.config_dir" => self.api_module_config_dir(method, p),
+            "module.pane.open" => self.api_module_pane_open(method, p),
             // ── module settings (docs/13 §3.6) ──
-            "module.settings.list" => {
-                let id = req_str(p, "id")?.to_string();
-                let values = self.module_settings(&id).map_err(module_err)?;
-                let specs: Vec<Value> = self
-                    .modules
-                    .find(&id)
-                    .map(|m| {
-                        m.manifest
-                            .settings
-                            .iter()
-                            .map(|s| {
-                                let v = values.get(&s.key).cloned().unwrap_or(Value::Null);
-                                // A listing is the "show me everything" call and
-                                // usually lands in a terminal, so a secret reports
-                                // only whether it is set — same as the UI. Read the
-                                // exact value with `module.settings.get {key}`.
-                                let set = !matches!(&v, Value::Null)
-                                    && !v.as_str().is_some_and(|t| t.is_empty());
-                                json!({
-                                    "key": s.key, "title": s.title, "type": s.kind,
-                                    "options": s.options, "min": s.min, "max": s.max,
-                                    "secret": s.secret, "set": set,
-                                    "value": if s.secret { Value::Null } else { v },
-                                })
-                            })
-                            .collect()
-                    })
-                    .unwrap_or_default();
-                Ok(json!({"type":"module_settings","id": id,"settings": specs}))
-            }
-            "module.settings.get" => {
-                let id = req_str(p, "id")?.to_string();
-                let values = self.module_settings(&id).map_err(module_err)?;
-                match p.get("key").and_then(|v| v.as_str()) {
-                    Some(k) => {
-                        let v = values
-                            .get(k)
-                            .cloned()
-                            .ok_or_else(|| module_err(format!("module {id} has no setting {k}")))?;
-                        Ok(json!({"type":"module_setting","id": id,"key": k,"value": v}))
-                    }
-                    None => Ok(json!({"type":"module_settings","id": id,"values": values})),
-                }
-            }
-            "module.settings.set" => {
-                let id = req_str(p, "id")?.to_string();
-                let key = req_str(p, "key")?.to_string();
-                // Accept a JSON value or a bare string (what the CLI sends).
-                let raw = p.get("value").cloned().unwrap_or(Value::Null);
-                let v = self
-                    .module_set_setting(&id, &key, raw)
-                    .map_err(module_err)?;
-                Ok(json!({"type":"module_setting","id": id,"key": key,"value": v}))
-            }
-            "module.pane.focus" => {
-                let id = self.resolve_pane(p)?.ok_or_else(not_found)?;
-                self.focus_pane_global(id);
-                Ok(json!({"type":"ok"}))
-            }
-            "module.pane.close" => {
-                let id = self.resolve_pane(p)?.ok_or_else(not_found)?;
-                self.close_pane(id);
-                Ok(json!({"type":"ok"}))
-            }
+            "module.settings.list" => self.api_module_settings_list(method, p),
+            "module.settings.get" => self.api_module_settings_get(method, p),
+            "module.settings.set" => self.api_module_settings_set(method, p),
+            "module.pane.focus" => self.api_module_pane_focus(method, p),
+            "module.pane.close" => self.api_module_pane_close(method, p),
             // ── DIFF review (docs/88) ────────────────────────────────────
-            "diff.refresh" => {
-                self.ensure_diff_snapshot().map_err(diff_err)?;
-                let generation = self
-                    .diff
-                    .snapshot
-                    .as_ref()
-                    .map(|snapshot| snapshot.generation)
-                    .unwrap_or_default();
-                Ok(json!({"type":"ok","refresh":"complete","generation":generation}))
-            }
-            "diff.list" => {
-                self.ensure_diff_snapshot().map_err(diff_err)?;
-                let layer = p
-                    .get("layer")
-                    .and_then(|value| value.as_str())
-                    .map(parse_diff_layer)
-                    .transpose()?;
-                let snapshot = self
-                    .diff
-                    .snapshot
-                    .as_ref()
-                    .ok_or_else(|| diff_err("DIFF is not ready".to_string()))?;
-                let files: Vec<Value> = snapshot
-                    .files
-                    .iter()
-                    .filter(|file| layer.as_ref().is_none_or(|layer| &file.key.layer == layer))
-                    .map(diff_file_json)
-                    .collect();
-                Ok(json!({
-                    "type":"diff_list",
-                    "repo": snapshot.repo_root,
-                    "branch": snapshot.branch,
-                    "generation": snapshot.generation,
-                    "fingerprint": snapshot.fingerprint,
-                    "omitted": snapshot.omitted_files,
-                    "refreshing": self.diff.status_inflight,
-                    "files": files,
-                }))
-            }
-            "diff.open" => {
-                self.ensure_diff_snapshot().map_err(diff_err)?;
-                let target = match p
-                    .get("placement")
-                    .or_else(|| p.get("target"))
-                    .and_then(|v| v.as_str())
-                {
-                    Some("tab") => crate::app::files::OpenTarget::Tab,
-                    Some("pane") => crate::app::files::OpenTarget::Pane,
-                    Some("preview") | None => crate::app::files::OpenTarget::Preview,
-                    Some(_) => {
-                        return Err(diff_err(
-                            "placement must be preview, pane, or tab".to_string(),
-                        ))
-                    }
-                };
-                let preference = match p.get("view").and_then(Value::as_str) {
-                    None => None,
-                    Some("auto") => Some(crate::diff::DiffLayoutPreference::Auto),
-                    Some("split") => Some(crate::diff::DiffLayoutPreference::Split),
-                    Some("stack") => Some(crate::diff::DiffLayoutPreference::Stack),
-                    Some(_) => {
-                        return Err(diff_err("view must be auto, split, or stack".to_string()))
-                    }
-                };
-                let layer = p
-                    .get("layer")
-                    .and_then(|value| value.as_str())
-                    .map(parse_diff_layer)
-                    .transpose()?;
-                let raw = p.get("path").and_then(|value| value.as_str()).unwrap_or("");
-                let key = if raw.is_empty() {
-                    self.diff
-                        .selected_file()
-                        .or_else(|| {
-                            self.diff
-                                .snapshot
-                                .as_ref()
-                                .and_then(|snapshot| snapshot.files.first())
-                        })
-                        .map(|file| file.key.clone())
-                        .ok_or_else(|| diff_err("there are no changed files".to_string()))?
-                } else {
-                    self.diff_file_for_path(raw, layer.as_ref())
-                        .map_err(diff_err)?
-                        .key
-                };
-                self.open_diff_view(key.clone(), target);
-                let id = self
-                    .diff_view_showing(&key)
-                    .ok_or_else(|| diff_err("failed to open diff".to_string()))?;
-                if let (Some(preference), Some(crate::app::ViewKind::Diff(view))) =
-                    (preference, self.views.get_mut(&id))
-                {
-                    view.preference = preference;
-                }
-                Ok(
-                    json!({"type":"diff_open","pane":id.0.to_string(),"path":key.display_path(),"layer":key.layer.label()}),
-                )
-            }
-            "diff.get" => {
-                self.ensure_diff_snapshot().map_err(diff_err)?;
-                let raw = req_str(p, "path")?;
-                let layer = p
-                    .get("layer")
-                    .and_then(|value| value.as_str())
-                    .map(parse_diff_layer)
-                    .transpose()?;
-                let file = self
-                    .diff_file_for_path(raw, layer.as_ref())
-                    .map_err(diff_err)?;
-                let diff = self.load_diff_file_sync(&file).map_err(diff_err)?;
-                let include_patch = p
-                    .get("include_patch")
-                    .and_then(Value::as_bool)
-                    .unwrap_or(false);
-                let hunks: Vec<Value> = diff
-                    .hunks
-                    .iter()
-                    .map(|hunk| {
-                        json!({
-                            "id":hunk.id,
-                            "old_start":hunk.old_start,
-                            "new_start":hunk.new_start,
-                            "header":hunk.header,
-                            "lines": if include_patch {
-                                serde_json::to_value(&hunk.lines).unwrap_or(Value::Null)
-                            } else {
-                                Value::Null
-                            },
-                        })
-                    })
-                    .collect();
-                Ok(json!({
-                    "type":"diff",
-                    "file":diff_file_json(&file),
-                    "additions":diff.additions,
-                    "deletions":diff.deletions,
-                    "binary":diff.binary,
-                    "truncated":diff.truncated,
-                    "omitted_lines":diff.omitted_lines,
-                    "hunks":hunks,
-                }))
-            }
-            "diff.navigate" => {
-                let id = self.resolve_pane_or_focus(p)?;
-                let action = req_str(p, "action")?;
-                let key = match action {
-                    "next" | "next_line" => KeyEvent::new(KeyCode::Down, KeyModifiers::NONE),
-                    "previous" | "previous_line" => KeyEvent::new(KeyCode::Up, KeyModifiers::NONE),
-                    "next_file" => KeyEvent::new(KeyCode::Char('J'), KeyModifiers::NONE),
-                    "previous_file" => KeyEvent::new(KeyCode::Char('K'), KeyModifiers::NONE),
-                    "next_hunk" => KeyEvent::new(KeyCode::Char('}'), KeyModifiers::NONE),
-                    "previous_hunk" => KeyEvent::new(KeyCode::Char('{'), KeyModifiers::NONE),
-                    "next_note" => KeyEvent::new(KeyCode::Char('N'), KeyModifiers::NONE),
-                    "previous_note" => KeyEvent::new(KeyCode::Char('P'), KeyModifiers::NONE),
-                    "top" => KeyEvent::new(KeyCode::Home, KeyModifiers::NONE),
-                    "bottom" => KeyEvent::new(KeyCode::End, KeyModifiers::NONE),
-                    _ => {
-                        return Err(diff_err(
-                            "action must target a line, file, hunk, note, top, or bottom"
-                                .to_string(),
-                        ))
-                    }
-                };
-                if !self.handle_diff_key(id, key) {
-                    return Err(diff_err("target is not an open DIFF view".to_string()));
-                }
-                Ok(json!({"type":"ok","pane":id.0.to_string()}))
-            }
-            "diff.note.list" => {
-                self.ensure_diff_snapshot().map_err(diff_err)?;
-                self.ensure_diff_notes_sync().map_err(diff_err)?;
-                let state = p
-                    .get("state")
-                    .and_then(Value::as_str)
-                    .map(parse_note_state)
-                    .transpose()?;
-                let path = p
-                    .get("file")
-                    .or_else(|| p.get("path"))
-                    .and_then(Value::as_str);
-                let notes: Vec<Value> = self
-                    .diff
-                    .notes
-                    .iter()
-                    .filter(|note| state.is_none_or(|state| note.state == state))
-                    .filter(|note| {
-                        path.is_none_or(|path| note.anchor.diff_key.display_path() == path)
-                    })
-                    .map(note_json)
-                    .collect();
-                Ok(json!({"type":"diff_notes","notes":notes}))
-            }
-            "diff.note.apply" => {
-                self.ensure_diff_snapshot().map_err(diff_err)?;
-                self.ensure_diff_notes_sync().map_err(diff_err)?;
-                let items = p
-                    .get("notes")
-                    .and_then(Value::as_array)
-                    .filter(|items| !items.is_empty())
-                    .ok_or_else(|| diff_err("notes must be a non-empty array".to_string()))?;
-                if self.diff.notes.len().saturating_add(items.len()) > crate::diff::NOTE_CAP {
-                    return Err(diff_err(format!(
-                        "review note limit is {}",
-                        crate::diff::NOTE_CAP
-                    )));
-                }
-                let mut notes = Vec::with_capacity(items.len());
-                for item in items {
-                    let raw = item
-                        .get("file")
-                        .or_else(|| item.get("path"))
-                        .and_then(Value::as_str)
-                        .ok_or_else(|| diff_err("every note needs a file".to_string()))?;
-                    let layer = item
-                        .get("layer")
-                        .and_then(Value::as_str)
-                        .map(parse_diff_layer)
-                        .transpose()?;
-                    let file = self
-                        .diff_file_for_path(raw, layer.as_ref())
-                        .map_err(diff_err)?;
-                    let old = diff_line_param(item, "old_line")?;
-                    let new = diff_line_param(item, "new_line")?;
-                    let (side, start) = match (old, new) {
-                        (Some(line), None) => (crate::diff::DiffSide::Old, line),
-                        (None, Some(line)) => (crate::diff::DiffSide::New, line),
-                        _ => {
-                            return Err(diff_err(
-                                "every note needs exactly one old_line or new_line".to_string(),
-                            ))
-                        }
-                    };
-                    let end = diff_line_param(item, "end_line")?.unwrap_or(start);
-                    let diff = self.load_diff_file_sync(&file).map_err(diff_err)?;
-                    let context = crate::diff::notes::anchor_context(&diff, side, start, end)
-                        .map_err(diff_err)?;
-                    let context_sha256 = crate::diff::notes::context_hash(&context);
-                    let context: String = context.chars().take(512).collect();
-                    let body = item
-                        .get("body")
-                        .and_then(Value::as_str)
-                        .ok_or_else(|| diff_err("every note needs a body".to_string()))?
-                        .to_string();
-                    let kind = parse_note_kind(
-                        item.get("kind").and_then(Value::as_str).unwrap_or("issue"),
-                    )?;
-                    let key = file.key;
-                    let now = crate::diff::notes::now_ms();
-                    notes.push(crate::diff::ReviewNote {
-                        id: crate::diff::notes::note_id(),
-                        review_id: crate::diff::notes::review_id(&key),
-                        author: "external".to_string(),
-                        kind,
-                        body,
-                        anchor: crate::diff::notes::NoteAnchor {
-                            diff_key: key,
-                            side,
-                            start_line: start,
-                            end_line: end,
-                            context_sha256,
-                            context,
-                        },
-                        state: crate::diff::NoteState::Open,
-                        deliveries: Vec::new(),
-                        revision: 1,
-                        created_at_ms: now,
-                        updated_at_ms: now,
-                    });
-                }
-                crate::diff::notes::save_batch_new(&notes).map_err(diff_err)?;
-                self.diff.notes.extend(notes.iter().cloned());
-                self.refresh_diff_note_counts();
-                Ok(json!({
-                    "type":"diff_notes_applied",
-                    "notes":notes.iter().map(note_json).collect::<Vec<_>>()
-                }))
-            }
-            "diff.note.add" => {
-                self.ensure_diff_snapshot().map_err(diff_err)?;
-                self.ensure_diff_notes_sync().map_err(diff_err)?;
-                let raw = p
-                    .get("file")
-                    .or_else(|| p.get("path"))
-                    .and_then(Value::as_str)
-                    .ok_or_else(|| diff_err("file is required".to_string()))?;
-                let layer = p
-                    .get("layer")
-                    .and_then(Value::as_str)
-                    .map(parse_diff_layer)
-                    .transpose()?;
-                let file = self
-                    .diff_file_for_path(raw, layer.as_ref())
-                    .map_err(diff_err)?;
-                let (side, start) = match (
-                    diff_line_param(p, "old_line")?,
-                    diff_line_param(p, "new_line")?,
-                ) {
-                    (Some(line), None) => (crate::diff::DiffSide::Old, line),
-                    (None, Some(line)) => (crate::diff::DiffSide::New, line),
-                    _ => {
-                        return Err(diff_err(
-                            "pass exactly one of old_line or new_line".to_string(),
-                        ))
-                    }
-                };
-                let end = diff_line_param(p, "end_line")?.unwrap_or(start);
-                let diff = self.load_diff_file_sync(&file).map_err(diff_err)?;
-                let context = crate::diff::notes::anchor_context(&diff, side, start, end)
-                    .map_err(diff_err)?;
-                let context_sha256 = crate::diff::notes::context_hash(&context);
-                let context: String = context.chars().take(512).collect();
-                let body = req_str(p, "body")?.to_string();
-                let kind =
-                    parse_note_kind(p.get("kind").and_then(Value::as_str).unwrap_or("issue"))?;
-                let key = file.key;
-                let now = crate::diff::notes::now_ms();
-                let note = crate::diff::ReviewNote {
-                    id: crate::diff::notes::note_id(),
-                    review_id: crate::diff::notes::review_id(&key),
-                    author: "external".to_string(),
-                    kind,
-                    body,
-                    anchor: crate::diff::notes::NoteAnchor {
-                        diff_key: key,
-                        side,
-                        start_line: start,
-                        end_line: end,
-                        context_sha256,
-                        context,
-                    },
-                    state: crate::diff::NoteState::Open,
-                    deliveries: Vec::new(),
-                    revision: 1,
-                    created_at_ms: now,
-                    updated_at_ms: now,
-                };
-                crate::diff::notes::save(&note, None).map_err(diff_err)?;
-                self.apply_diff_note_saved(note.clone(), Ok(()));
-                Ok(json!({"type":"diff_note","note":note_json(&note)}))
-            }
+            "diff.refresh" => self.api_diff_refresh(method, p),
+            "diff.list" => self.api_diff_list(method, p),
+            "diff.open" => self.api_diff_open(method, p),
+            "diff.get" => self.api_diff_get(method, p),
+            "diff.navigate" => self.api_diff_navigate(method, p),
+            "diff.note.list" => self.api_diff_note_list(method, p),
+            "diff.note.apply" => self.api_diff_note_apply(method, p),
+            "diff.note.add" => self.api_diff_note_add(method, p),
             "diff.note.edit" | "diff.note.resolve" | "diff.note.reopen" => {
-                self.ensure_diff_snapshot().map_err(diff_err)?;
-                self.ensure_diff_notes_sync().map_err(diff_err)?;
-                let id = req_str(p, "id")?;
-                let note = self
-                    .diff
-                    .notes
-                    .iter()
-                    .find(|note| note.id == id)
-                    .cloned()
-                    .ok_or_else(|| diff_err("review note not found".to_string()))?;
-                let mut updated = note.clone();
-                match method {
-                    "diff.note.edit" => updated.body = req_str(p, "body")?.to_string(),
-                    "diff.note.resolve" => updated.state = crate::diff::NoteState::Resolved,
-                    "diff.note.reopen" => updated.state = crate::diff::NoteState::Open,
-                    _ => unreachable!(),
-                }
-                updated.revision = updated.revision.saturating_add(1);
-                updated.updated_at_ms = crate::diff::notes::now_ms();
-                crate::diff::notes::save(&updated, Some(note.revision)).map_err(diff_err)?;
-                self.apply_diff_note_saved(updated.clone(), Ok(()));
-                Ok(json!({"type":"diff_note","note":note_json(&updated)}))
+                self.api_diff_note_edit(method, p)
             }
-            "diff.note.remove" => {
-                self.ensure_diff_snapshot().map_err(diff_err)?;
-                self.ensure_diff_notes_sync().map_err(diff_err)?;
-                let id = req_str(p, "id")?;
-                let note = self
-                    .diff
-                    .notes
-                    .iter()
-                    .find(|note| note.id == id)
-                    .cloned()
-                    .ok_or_else(|| diff_err("review note not found".to_string()))?;
-                crate::diff::notes::remove(&note, Some(note.revision)).map_err(diff_err)?;
-                self.apply_diff_note_removed(id.to_string(), Ok(()));
-                Ok(json!({"type":"ok","removed":id}))
-            }
-            "diff.note.send" => {
-                self.ensure_diff_snapshot().map_err(diff_err)?;
-                self.ensure_diff_notes_sync().map_err(diff_err)?;
-                let target = req_str(p, "to")?;
-                let all_open = p.get("all_open").and_then(Value::as_bool).unwrap_or(false);
-                let ids: Vec<&str> = p
-                    .get("ids")
-                    .and_then(Value::as_array)
-                    .map(|items| items.iter().filter_map(Value::as_str).collect())
-                    .unwrap_or_default();
-                let selected: Vec<crate::diff::ReviewNote> = self
-                    .diff
-                    .notes
-                    .iter()
-                    .filter(|note| {
-                        if all_open {
-                            note.state == crate::diff::NoteState::Open
-                        } else {
-                            ids.contains(&note.id.as_str())
-                        }
-                    })
-                    .cloned()
-                    .collect();
-                if selected.is_empty() {
-                    return Err(diff_err("select at least one review note".to_string()));
-                }
-                let params = json!({"target":target});
-                let pane_id = self.resolve_agent_target(&params)?;
-                let selected_ids: Vec<String> =
-                    selected.iter().map(|note| note.id.clone()).collect();
-                let count = self
-                    .deliver_diff_notes(pane_id, target, &selected_ids)
-                    .map_err(diff_err)?;
-                Ok(
-                    json!({"type":"diff_note_send","pane":pane_id.0.to_string(),"target":target,"count":count}),
-                )
-            }
+            "diff.note.remove" => self.api_diff_note_remove(method, p),
+            "diff.note.send" => self.api_diff_note_send(method, p),
             // ── git (docs/17) — fast local-git reads + open the git tab ──
-            "git.status" => {
-                let cwd = self.git_workspace_cwd(p);
-                let s = crate::git::local::status(&cwd).map_err(git_err)?;
-                let files = |v: &[crate::git::model::FileChange]| -> Vec<Value> {
-                    v.iter()
-                        .map(|c| json!({"code": c.code.to_string(), "path": c.path}))
-                        .collect()
-                };
-                Ok(json!({
-                    "type": "git_status", "branch": s.branch, "upstream": s.upstream,
-                    "ahead": s.ahead, "behind": s.behind,
-                    "staged": files(&s.staged), "unstaged": files(&s.unstaged),
-                    "untracked": s.untracked, "stashes": s.stashes,
-                }))
-            }
-            "git.branches" => {
-                let cwd = self.git_workspace_cwd(p);
-                let v = crate::git::local::branches(&cwd).map_err(git_err)?;
-                let arr: Vec<Value> = v
-                    .iter()
-                    .map(|b| json!({"name": b.name, "head": b.is_head, "ahead": b.ahead, "behind": b.behind, "subject": b.subject}))
-                    .collect();
-                Ok(json!({"type":"git_branches","branches":arr}))
-            }
-            "git.log" => {
-                let cwd = self.git_workspace_cwd(p);
-                let n = param_usize(p, "n").unwrap_or(30);
-                let v = crate::git::local::commits(&cwd, n, false).map_err(git_err)?;
-                let arr: Vec<Value> = v
-                    .iter()
-                    .map(|c| json!({"sha": c.sha, "subject": c.subject, "author": c.author, "when": c.when, "refs": c.refs}))
-                    .collect();
-                Ok(json!({"type":"git_log","commits":arr}))
-            }
-            "git.open" => {
-                let i = param_usize(p, "workspace")
-                    .or_else(|| param_usize(p, "node"))
-                    .unwrap_or(self.active_ws);
-                self.open_git_tab(i);
-                Ok(json!({"type":"ok","git": self.active_is_git()}))
-            }
-            "mission.snapshot" | "mission.refresh" => {
-                reject_api_fields(p, &["scope", "workspace", "workspace_id"])?;
-                let scope = match p.get("scope") {
-                    None => crate::mission::MissionScope::Workspace,
-                    Some(Value::String(scope)) if scope == "workspace" => {
-                        crate::mission::MissionScope::Workspace
-                    }
-                    Some(Value::String(scope)) if scope == "all" => {
-                        crate::mission::MissionScope::All
-                    }
-                    Some(_) => {
-                        return Err((
-                            "invalid_request".to_string(),
-                            "scope must be workspace or all".to_string(),
-                        ))
-                    }
-                };
-                let workspace = self.optional_socket_workspace(p)?.unwrap_or(self.active_ws);
-                if scope == crate::mission::MissionScope::Workspace
-                    && workspace >= self.workspaces.len()
-                {
-                    return Err(workspace_update_error(
-                        workspace,
-                        WorkspaceUpdateError::NotFound,
-                    ));
-                }
-                if method == "mission.refresh" {
-                    self.request_mission_usage_refresh_for(scope, workspace);
-                    Ok(json!({
-                        "type":"mission_refresh",
-                        "scope":match scope { crate::mission::MissionScope::Workspace => "workspace", crate::mission::MissionScope::All => "all" },
-                        "workspace":workspace.to_string(),
-                        "refreshing":true,
-                    }))
-                } else {
-                    Ok(self.mission_snapshot_value(scope, workspace))
-                }
-            }
-            "mission.open" => {
-                let i = self.optional_socket_workspace(p)?.unwrap_or(self.active_ws);
-                if i >= self.workspaces.len() {
-                    return Err(workspace_update_error(i, WorkspaceUpdateError::NotFound));
-                }
-                self.open_mission_control(i);
-                Ok(json!({"type":"ok","mission": self.active_is_mission()}))
-            }
+            "git.status" => self.api_git_status(method, p),
+            "git.branches" => self.api_git_branches(method, p),
+            "git.log" => self.api_git_log(method, p),
+            "git.open" => self.api_git_open(method, p),
+            "mission.snapshot" | "mission.refresh" => self.api_mission_snapshot(method, p),
+            "mission.open" => self.api_mission_open(method, p),
             // ── file viewer (docs/38) ──
-            "files.open" => {
-                let raw = p.get("path").and_then(|v| v.as_str()).unwrap_or("");
-                if raw.is_empty() {
-                    return Err(("bad_request".into(), "path required".into()));
-                }
-                let path = self.resolve_file_path(raw);
-                let target = match p.get("target").and_then(|v| v.as_str()) {
-                    Some("tab") => crate::app::files::OpenTarget::Tab,
-                    Some("pane") => crate::app::files::OpenTarget::Pane,
-                    _ => crate::app::files::OpenTarget::Preview,
-                };
-                self.open_file_view(path, target);
-                Ok(json!({"type":"ok"}))
-            }
-            "files.tree" => {
-                self.prepare_file_tree_api(false);
-                let rows: Vec<Value> = self
-                    .file_tree
-                    .visible_rows()
-                    .iter()
-                    .map(|r| {
-                        json!({
-                            "path": r.path.to_string_lossy(),
-                            "name": r.name,
-                            "depth": r.depth,
-                            "dir": r.is_dir,
-                            "expanded": r.expanded,
-                        })
-                    })
-                    .collect();
-                Ok(json!({
-                    "type": "file_tree",
-                    "root": self.file_tree.root().to_string_lossy(),
-                    "rows": rows,
-                }))
-            }
-            "files.reveal" => {
-                let raw = p.get("path").and_then(|v| v.as_str()).unwrap_or("");
-                if raw.is_empty() {
-                    return Err(("bad_request".into(), "path required".into()));
-                }
-                let path = self.resolve_file_path(raw);
-                self.file_tree.reveal(&path);
-                Ok(json!({"type":"ok"}))
-            }
-            "files.refresh" => {
-                self.prepare_file_tree_api(true);
-                Ok(json!({"type":"ok"}))
-            }
+            "files.open" => self.api_files_open(method, p),
+            "files.tree" => self.api_files_tree(method, p),
+            "files.reveal" => self.api_files_reveal(method, p),
+            "files.refresh" => self.api_files_refresh(method, p),
             // ── worktrees (docs/18 WT-3) ──
-            "worktree.list" => {
-                let cwd = self.git_workspace_cwd(p);
-                let v = crate::git::local::worktrees(&cwd).map_err(git_err)?;
-                let arr: Vec<Value> = v
-                    .iter()
-                    .map(|w| {
-                        json!({"path": w.path.display().to_string(), "branch": w.branch, "head": w.head, "main": w.is_main})
-                    })
-                    .collect();
-                Ok(json!({"type":"worktree_list","worktrees":arr}))
-            }
-            "worktree.create" => {
-                let branch = p.get("branch").and_then(|v| v.as_str()).unwrap_or("");
-                let repo = self.git_workspace_cwd(p);
-                let path = self.create_worktree(&repo, branch).map_err(git_err)?;
-                Ok(json!({"type":"ok","path": path.display().to_string()}))
-            }
-            "worktree.open" => {
-                let path = param_path(p)?;
-                if !self.create_workspace_at(path.clone()) {
-                    return Err((
-                        "spawn_failed".to_string(),
-                        format!(
-                            "couldn't open {} — the shell failed to start there",
-                            path.display()
-                        ),
-                    ));
-                }
-                Ok(json!({"type":"ok"}))
-            }
-            "worktree.remove" => {
-                let path = param_path(p)?;
-                // Run from the repo's **main** worktree — git refuses to remove a
-                // worktree from inside it, and the active workspace may be unrelated.
-                let repo = crate::git::local::worktrees(&path)
-                    .ok()
-                    .and_then(|wts| wts.into_iter().find(|w| w.is_main).map(|w| w.path))
-                    .unwrap_or_else(|| self.ws().cwd.clone());
-                crate::git::local::worktree_remove(&repo, &path).map_err(git_err)?;
-                // Tidy the now-possibly-empty `worktrees/<repo>/` parent — but only
-                // under our managed dir, and `remove_dir` only succeeds if empty.
-                if let Some(parent) = path.parent() {
-                    if parent.starts_with(crate::persist::config_dir().join("worktrees")) {
-                        let _ = std::fs::remove_dir(parent);
-                    }
-                }
-                // Close the workspace opened at this worktree, if any.
-                if let Some(i) = self
-                    .workspaces
-                    .iter()
-                    .position(|w| crate::platform::same_path(&w.cwd, &path))
-                {
-                    self.close_workspace(i);
-                }
-                Ok(json!({"type":"ok"}))
-            }
+            "worktree.list" => self.api_worktree_list(method, p),
+            "worktree.create" => self.api_worktree_create(method, p),
+            "worktree.open" => self.api_worktree_open(method, p),
+            "worktree.remove" => self.api_worktree_remove(method, p),
             // ── Agent Automation (docs/118): durable schedules over ORCH ───
-            "automation.create" | "automation.update" => {
-                reject_api_fields(
-                    p,
-                    if method == "automation.create" {
-                        &[
-                            "name",
-                            "enabled",
-                            "trigger",
-                            "target",
-                            "task",
-                            "policy",
-                            "idempotency_key",
-                        ]
-                    } else {
-                        &[
-                            "id", "name", "enabled", "trigger", "target", "task", "policy",
-                        ]
-                    },
-                )?;
-                let now = crate::automation::unix_now();
-                let mut input = automation_input(p)?;
-                if method == "automation.create" {
-                    if let Some(descriptor) = crate::agent::registry::find(&input.task.agent_id) {
-                        input.task.agent_id = descriptor.id.to_string();
-                    }
-                    if let Some(automation) = self
-                        .automation
-                        .create_retry(&input, opt_borrowed_str(p, "idempotency_key"))
-                        .map_err(automation_err)?
-                    {
-                        let state = self.durable_active_target_state(&automation);
-                        return Ok(
-                            json!({"type":"automation", "automation":crate::automation::public_automation(&automation, state)}),
-                        );
-                    }
-                }
-                validate_automation_target(self, &mut input)?;
-                let automation = if method == "automation.create" {
-                    self.automation
-                        .create(input, opt_borrowed_str(p, "idempotency_key"), now)
-                        .map_err(automation_err)?
-                } else {
-                    let id = req_str(p, "id")?;
-                    self.automation
-                        .update(id, input, now)
-                        .map_err(automation_err)?
-                };
-                self.persist_automation();
-                if automation.target.is_durable_active_agent() {
-                    self.initialize_durable_active_target_state(&automation);
-                } else {
-                    self.automation.ready_active_targets.remove(&automation.id);
-                    self.automation.active_target_states.remove(&automation.id);
-                }
-                self.emit_event(
-                    if method == "automation.create" {
-                        "automation.created"
-                    } else {
-                        "automation.updated"
-                    },
-                    crate::automation::definition_event(&automation),
-                );
-                let state = self.durable_active_target_state(&automation);
-                Ok(
-                    json!({"type":"automation", "automation":crate::automation::public_automation(&automation, state)}),
-                )
-            }
-            "automation.list" => {
-                reject_api_fields(p, &[])?;
-                let automations = self
-                    .automation
-                    .automations
-                    .iter()
-                    .map(|automation| {
-                        crate::automation::public_automation(
-                            automation,
-                            self.durable_active_target_state(automation),
-                        )
-                    })
-                    .collect::<Vec<_>>();
-                Ok(json!({
-                    "type":"automation_list",
-                    "automations": automations,
-                }))
-            }
-            "automation.get" => {
-                reject_api_fields(p, &["id"])?;
-                let id = req_str(p, "id")?;
-                let automation = self
-                    .automation
-                    .automation(id)
-                    .ok_or_else(|| ("not_found".into(), format!("no such automation: {id}")))?;
-                Ok(
-                    json!({"type":"automation", "automation":crate::automation::public_automation(automation, self.durable_active_target_state(automation))}),
-                )
-            }
-            "automation.enable" | "automation.disable" => {
-                reject_api_fields(p, &["id"])?;
-                let id = req_str(p, "id")?;
-                let enable = method == "automation.enable";
-                if enable {
-                    let automation =
-                        self.automation.automation(id).cloned().ok_or_else(|| {
-                            ("not_found".into(), format!("no such automation: {id}"))
-                        })?;
-                    if matches!(
-                        automation.target,
-                        crate::automation::AutomationTarget::ActiveAgent { .. }
-                    ) {
-                        self.validate_active_agent_target(&automation.target, &automation.task)?;
-                    }
-                }
-                let automation = self
-                    .automation
-                    .set_enabled(id, enable, crate::automation::unix_now())
-                    .map_err(automation_err)?;
-                self.persist_automation();
-                self.emit_event(
-                    if automation.enabled {
-                        "automation.enabled"
-                    } else {
-                        "automation.disabled"
-                    },
-                    crate::automation::definition_event(&automation),
-                );
-                let state = self.durable_active_target_state(&automation);
-                Ok(
-                    json!({"type":"automation", "automation":crate::automation::public_automation(&automation, state)}),
-                )
-            }
-            "automation.rebind" => {
-                reject_api_fields(p, &["id", "pane", "terminal_id"])?;
-                let id = req_str(p, "id")?.to_string();
-                if p.get("pane").is_none() {
-                    return Err((
-                        "invalid_request".into(),
-                        "automation.rebind requires pane".into(),
-                    ));
-                }
-                let pane = self.resolve_pane(p)?.ok_or_else(not_found)?;
-                let expected_terminal_id = optional_bounded_string(p, "terminal_id", 64)?;
-                let automation = self.rebind_active_agent_automation(
-                    &id,
-                    pane,
-                    expected_terminal_id.as_deref(),
-                )?;
-                let state = self.durable_active_target_state(&automation);
-                let event_state = self
-                    .automation
-                    .active_target_states
-                    .get(&automation.id)
-                    .copied()
-                    .unwrap_or(crate::automation::ActiveTargetState::NeedsRebind);
-                self.emit_event(
-                    "automation.rebound",
-                    crate::automation::definition_target_event(&automation, event_state),
-                );
-                Ok(json!({
-                    "type":"automation",
-                    "automation":crate::automation::public_automation(&automation, state),
-                }))
-            }
-            "automation.delete" => {
-                reject_api_fields(p, &["id"])?;
-                let id = req_str(p, "id")?;
-                let automation = self.automation.delete(id).map_err(automation_err)?;
-                self.persist_automation();
-                self.emit_event("automation.deleted", json!({"id":id}));
-                Ok(
-                    json!({"type":"automation", "automation":crate::automation::public_automation(&automation, None)}),
-                )
-            }
-            "automation.run" => {
-                reject_api_fields(p, &["id", "idempotency_key"])?;
-                if self.workspaces.is_empty() {
-                    return Err(("no_session".into(), "no active session".into()));
-                }
-                let id = req_str(p, "id")?.to_string();
-                let now = crate::automation::unix_now();
-                if let Some(run) = self
-                    .automation
-                    .run_retry(&id, opt_borrowed_str(p, "idempotency_key"))
-                    .map_err(automation_err)?
-                {
-                    return Ok(
-                        json!({"type":"automation_run", "run":crate::automation::public_run(&run)}),
-                    );
-                }
-                let run = self
-                    .automation
-                    .request_run(&id, opt_borrowed_str(p, "idempotency_key"), now)
-                    .map_err(automation_err)?;
-                self.persist_automation();
-                self.emit_event(
-                    "automation.run_queued",
-                    json!({"automation_id":run.automation_id, "run_id":run.id, "scheduled_at":run.scheduled_at}),
-                );
-                self.start_automation_run(&run.id, now);
-                let run = self.automation.run(&run.id).cloned().unwrap_or(run);
-                Ok(json!({"type":"automation_run", "run":crate::automation::public_run(&run)}))
-            }
-            "automation.history" => {
-                reject_api_fields(p, &["id", "limit"])?;
-                let id = opt_borrowed_str(p, "id");
-                let limit = p
-                    .get("limit")
-                    .and_then(Value::as_u64)
-                    .unwrap_or(50)
-                    .clamp(1, 200) as usize;
-                let runs = self
-                    .automation
-                    .runs
-                    .iter()
-                    .rev()
-                    .filter(|run| id.is_none_or(|id| run.automation_id == id))
-                    .take(limit)
-                    .map(crate::automation::public_run)
-                    .collect::<Vec<_>>();
-                Ok(json!({"type":"automation_history", "runs":runs}))
-            }
-            "automation.preview" => {
-                reject_api_fields(p, &["trigger", "from_utc"])?;
-                let trigger = automation_trigger(p.get("trigger"))?;
-                let now = p
-                    .get("from_utc")
-                    .and_then(Value::as_u64)
-                    .unwrap_or_else(crate::automation::unix_now);
-                let occurrences = crate::automation::AutomationState::preview(&trigger, now, 5)
-                    .map_err(automation_err)?;
-                Ok(json!({"type":"automation_preview", "occurrences_utc":occurrences}))
-            }
-            "automation.health" => {
-                reject_api_fields(p, &[])?;
-                Ok(json!({
-                    "type":"automation_health",
-                    "summary":self.automation.health(),
-                    "automations":self.automation_views(),
-                }))
-            }
+            "automation.create" | "automation.update" => self.api_automation_create(method, p),
+            "automation.list" => self.api_automation_list(method, p),
+            "automation.get" => self.api_automation_get(method, p),
+            "automation.enable" | "automation.disable" => self.api_automation_enable(method, p),
+            "automation.rebind" => self.api_automation_rebind(method, p),
+            "automation.delete" => self.api_automation_delete(method, p),
+            "automation.run" => self.api_automation_run(method, p),
+            "automation.history" => self.api_automation_history(method, p),
+            "automation.preview" => self.api_automation_preview(method, p),
+            "automation.health" => self.api_automation_health(method, p),
             // ── ORCH-1/2: task ledger + path leases (docs/22, M0) ──────────
-            "task.add" => {
-                reject_api_fields(p, &["title", "prompt", "paths", "deps", "gate"])?;
-                let title = req_str(p, "title")?.to_string();
-                let prompt = optional_task_prompt(p)?;
-                let task = self
-                    .orch
-                    .add_task_with_prompt(
-                        title,
-                        prompt,
-                        str_array(p, "paths"),
-                        str_array(p, "deps"),
-                        opt_str(p, "gate"),
-                    )
-                    .map_err(orch_err)?;
-                self.orch.save();
-                self.emit_event("task.added", task_json(&task));
-                Ok(json!({ "type": "task", "task": task_json(&task) }))
-            }
-            "task.list" => Ok(json!({
-                "type": "task_list",
-                "tasks": self.orch.tasks.iter().map(task_json).collect::<Vec<_>>(),
-            })),
-            "task.get" => {
-                let id = req_str(p, "id")?;
-                match self.orch.task(id) {
-                    Some(t) => Ok(json!({ "type": "task", "task": task_json(t) })),
-                    None => Err(("not_found".into(), format!("no such task: {id}"))),
-                }
-            }
-            "task.claim" => {
-                let id = req_str(p, "id")?.to_string();
-                let pane = self.orch_pane(p)?;
-                let task = self.orch.claim(&id, pane).map_err(orch_err)?;
-                self.orch.save();
-                self.emit_event("task.claimed", task_json(&task));
-                Ok(json!({ "type": "task", "task": task_json(&task) }))
-            }
-            "task.start" => {
-                let id = req_str(p, "id")?.to_string();
-                let mode =
-                    task_worker_mode(p, self.orch.task(&id).and_then(|task| task.worker_mode))?;
-                let started = self.task_start(
-                    &id,
-                    opt_str(p, "branch"),
-                    opt_str(p, "agent"),
-                    mode,
-                    opt_str(p, "workspace_id"),
-                )?;
-                let task = self.orch.task(&id).map(task_json).unwrap_or(Value::Null);
-                Ok(json!({
-                    "type": "task",
-                    "task": task,
-                    "pane": started.pane.0.to_string(),
-                    "mode": started.mode.as_str(),
-                    "workspace_id": started.workspace_id,
-                    "tab_id": started.tab_id,
-                    "cwd": started.cwd.display().to_string(),
-                    "worktree": started.worktree,
-                    "branch": started.branch,
-                }))
-            }
-            "task.update" => {
-                reject_api_fields(p, &["id", "status", "output", "note", "prompt"])?;
-                let id = req_str(p, "id")?.to_string();
-                let status = if let Some(s) = p.get("status").and_then(|v| v.as_str()) {
-                    let st = crate::orch::TaskStatus::parse(s).ok_or_else(|| {
-                        ("bad_request".to_string(), format!("unknown status: {s}"))
-                    })?;
-                    if matches!(
-                        st,
-                        crate::orch::TaskStatus::Merging | crate::orch::TaskStatus::Merged
-                    ) {
-                        return Err((
-                            "protected_status".to_string(),
-                            format!("{s} is set only by task.merge"),
-                        ));
-                    }
-                    Some(st)
-                } else {
-                    None
-                };
-                if let Some(current) = self.orch.task(&id).map(|task| task.status) {
-                    if matches!(
-                        current,
-                        crate::orch::TaskStatus::Merging | crate::orch::TaskStatus::Merged
-                    ) {
-                        return Err((
-                            "task_complete".to_string(),
-                            format!("{id} is already {}", current.as_str()),
-                        ));
-                    }
-                }
-                if p.get("prompt").is_some() {
-                    self.orch
-                        .set_prompt(&id, optional_task_prompt(p)?)
-                        .map_err(orch_err)?;
-                }
-                if let Some(st) = status {
-                    self.orch.set_status(&id, st).map_err(orch_err)?;
-                }
-                if let Some(o) = p.get("output").and_then(|v| v.as_str()) {
-                    self.orch.add_output(&id, o.to_string()).map_err(orch_err)?;
-                }
-                if let Some(n) = p.get("note").and_then(|v| v.as_str()) {
-                    self.orch.add_note(&id, n.to_string()).map_err(orch_err)?;
-                }
-                self.orch.save();
-                let t = self.orch.task(&id).cloned();
-                let jv = t.as_ref().map(task_json).unwrap_or(Value::Null);
-                self.emit_event("task.updated", jv.clone());
-                self.sync_automation_task(&id);
-                Ok(json!({ "type": "task", "task": jv }))
-            }
-            "task.done" => {
-                // ORCH-5: if the task has a quality gate, `complete_task` runs it
-                // async and holds the task at Running until it passes (→ Done, and
-                // dependents announced) or fails (→ Review). No gate → done now.
-                let id = req_str(p, "id")?.to_string();
-                let gate_running = self.complete_task(&id)?;
-                let task = self.orch.task(&id).map(task_json).unwrap_or(Value::Null);
-                Ok(json!({ "type": "task", "task": task, "gate_running": gate_running }))
-            }
-            "task.merge" => {
-                // Socket requests are parked by `handle_task_merge_request` so
-                // Git never runs on the app loop. Reaching direct dispatch is a
-                // programmer error, not a second synchronous implementation.
-                Err((
-                    "async_required".to_string(),
-                    "task.merge must run through the control API".to_string(),
-                ))
-            }
-            "task.next" => {
-                // ORCH-4 scheduler: hand out the next ready task. `--start`
-                // spawns the requested worker mode; otherwise claim it here.
-                match self.orch.next_ready() {
-                    None => Ok(json!({ "type": "none", "message": "no ready tasks" })),
-                    Some(id) => {
-                        if p.get("start").and_then(|v| v.as_bool()).unwrap_or(false) {
-                            let mode = task_worker_mode(
-                                p,
-                                self.orch.task(&id).and_then(|task| task.worker_mode),
-                            )?;
-                            let started = self.task_start(
-                                &id,
-                                None,
-                                opt_str(p, "agent"),
-                                mode,
-                                opt_str(p, "workspace_id"),
-                            )?;
-                            let task = self.orch.task(&id).map(task_json).unwrap_or(Value::Null);
-                            Ok(json!({
-                                "type": "task", "task": task,
-                                "pane": started.pane.0.to_string(),
-                                "mode": started.mode.as_str(),
-                                "workspace_id": started.workspace_id,
-                                "tab_id": started.tab_id,
-                                "cwd": started.cwd.display().to_string(),
-                                "worktree": started.worktree,
-                                "branch": started.branch,
-                            }))
-                        } else {
-                            let pane = self.orch_pane(p)?;
-                            let task = self.orch.claim(&id, pane).map_err(orch_err)?;
-                            self.orch.save();
-                            self.emit_event("task.claimed", task_json(&task));
-                            Ok(json!({ "type": "task", "task": task_json(&task) }))
-                        }
-                    }
-                }
-            }
-            "task.heartbeat" => {
-                // ORCH-5 compaction gate: a worker reports its context usage.
-                let id = req_str(p, "id")?.to_string();
-                let ctx = p.get("context").and_then(|v| v.as_f64()).ok_or_else(|| {
-                    (
-                        "invalid_request".to_string(),
-                        "context (0..1) is required".to_string(),
-                    )
-                })?;
-                if !ctx.is_finite() || !(0.0..=1.0).contains(&ctx) {
-                    return Err((
-                        "invalid_request".to_string(),
-                        "context must be a finite number from 0 to 1".to_string(),
-                    ));
-                }
-                let over = self.orch.heartbeat(&id, ctx).map_err(orch_err)?;
-                self.orch.save();
-                if over {
-                    self.emit_event("task.needs_compaction", json!({ "id": id, "context": ctx }));
-                }
-                Ok(json!({ "type": "ok", "over_threshold": over }))
-            }
-            "task.delete" => {
-                let id = req_str(p, "id")?.to_string();
-                let task = self.orch.delete_task(&id).map_err(orch_err)?;
-                self.orch.save();
-                self.emit_event("task.deleted", json!({ "id": id }));
-                Ok(json!({ "type": "task", "task": task_json(&task) }))
-            }
-            "task.release" => {
-                let id = req_str(p, "id")?.to_string();
-                let task = self.orch.release_task(&id).map_err(orch_err)?;
-                let released = self.orch.release_task_leases(&id);
-                self.orch.save();
-                self.emit_event("task.released", task_json(&task));
-                self.sync_automation_task(&id);
-                Ok(json!({ "type": "task", "task": task_json(&task), "released_leases": released }))
-            }
-            "lease.acquire" => {
-                let task = req_str(p, "task")?.to_string();
-                let pane = self.orch_pane(p)?;
-                let lease = self
-                    .orch
-                    .acquire_lease(pane, task, str_array(p, "paths"))
-                    .map_err(orch_err)?;
-                self.orch.save();
-                self.emit_event(
-                    "lease.acquired",
-                    serde_json::to_value(&lease).unwrap_or(Value::Null),
-                );
-                Ok(
-                    json!({ "type": "lease", "lease": serde_json::to_value(&lease).unwrap_or(Value::Null) }),
-                )
-            }
-            "lease.release" => {
-                let id = req_str(p, "id")?;
-                self.orch.release_lease(id).map_err(orch_err)?;
-                self.orch.save();
-                self.emit_event("lease.released", json!({ "id": id }));
-                Ok(json!({ "type": "ok" }))
-            }
-            "lease.list" => Ok(json!({
-                "type": "lease_list",
-                "leases": serde_json::to_value(&self.orch.leases).unwrap_or(Value::Null),
-            })),
+            "task.add" => self.api_task_add(method, p),
+            "task.list" => self.api_task_list(method, p),
+            "task.get" => self.api_task_get(method, p),
+            "task.claim" => self.api_task_claim(method, p),
+            "task.start" => self.api_task_start(method, p),
+            "task.update" => self.api_task_update(method, p),
+            "task.done" => self.api_task_done(method, p),
+            "task.merge" => self.api_task_merge(method, p),
+            "task.next" => self.api_task_next(method, p),
+            "task.heartbeat" => self.api_task_heartbeat(method, p),
+            "task.delete" => self.api_task_delete(method, p),
+            "task.release" => self.api_task_release(method, p),
+            "lease.acquire" => self.api_lease_acquire(method, p),
+            "lease.release" => self.api_lease_release(method, p),
+            "lease.list" => self.api_lease_list(method, p),
             other => Err((
                 "invalid_request".to_string(),
                 format!("unknown method: {other}"),
